chore: us new TS config, move all the src files to new src dir [PT-188049278]
