--- js/lib.ts
@@ -1,3 +1,0 @@
-import Seasons from './components/seasons';
-
-exports.Seasons = Seasons;
