--- js/models/museo-500-regular.ts
@@ -1,2 +1,0 @@
-var fontDef = {"glyphs":{"0":{"ha":843,"x_min":0,"x_max":0,"o":"m 422 -17 b 75 492 160 -17 75 208 b 422 997 75 774 160 997 b 768 492 683 997 768 774 b 422 -17 768 208 683 -17 m 422 110 b 628 492 582 110 628 278 b 422 871 628 704 582 871 b 215 492 261 871 215 704 b 422 110 215 278 261 110 "},"1":{"ha":700,"x_min":0,"x_max":0,"o":"m 89 119 l 304 119 l 304 753 b 306 814 304 783 306 814 l 303 814 b 257 757 303 814 292 792 l 165 667 l 82 753 l 317 981 l 443 981 l 443 119 l 657 119 l 657 0 l 89 0 "},"2":{"ha":796,"x_min":0,"x_max":0,"o":"m 53 113 b 551 714 53 474 551 449 b 379 869 551 804 485 869 b 228 776 292 869 228 825 l 228 733 l 100 733 l 100 817 b 385 997 100 944 271 997 b 693 715 564 997 693 883 b 196 146 693 378 196 372 b 231 119 196 126 207 119 l 579 119 b 610 150 599 119 610 131 l 610 213 l 738 213 l 738 89 b 649 0 738 26 711 0 l 146 0 b 53 113 69 0 53 33 "},"3":{"ha":751,"x_min":0,"x_max":0,"o":"m 111 225 b 360 114 111 225 208 114 b 554 293 465 114 554 188 b 343 468 554 403 457 468 l 274 468 l 244 542 l 449 799 b 506 863 478 840 506 863 l 506 865 b 422 861 506 865 476 861 l 233 861 b 203 831 214 861 203 851 l 203 768 l 75 768 l 75 892 b 160 981 75 954 97 981 l 676 981 l 676 893 l 418 582 b 696 294 546 568 696 483 b 360 -17 696 131 564 -17 b 36 121 156 -17 36 121 "},"4":{"ha":799,"x_min":0,"x_max":0,"o":"m 24 349 l 465 981 l 624 981 l 624 379 l 776 379 l 776 260 l 624 260 l 624 0 l 485 0 l 485 260 l 24 260 m 486 717 b 492 817 486 761 492 817 l 489 817 b 439 735 489 817 464 769 l 225 440 b 174 378 200 406 174 378 l 174 375 b 257 379 174 375 214 379 l 486 379 "},"5":{"ha":808,"x_min":0,"x_max":0,"o":"m 146 222 b 375 114 146 222 225 114 b 586 310 499 114 586 200 b 353 510 586 432 481 510 b 167 460 244 510 167 460 l 110 486 l 149 892 b 238 981 154 954 175 981 l 579 981 b 668 892 642 981 668 954 l 668 768 l 542 768 l 542 831 b 511 861 542 851 531 861 l 306 861 b 275 831 286 861 278 851 l 256 651 b 249 600 253 621 249 600 l 251 600 b 378 629 251 600 304 629 b 728 307 589 629 728 486 b 375 -17 728 135 589 -17 b 71 121 172 -17 71 121 "},"6":{"ha":788,"x_min":0,"x_max":0,"o":"m 419 -17 b 53 435 211 -17 53 161 b 494 997 53 690 190 997 b 685 957 611 997 685 957 l 640 839 b 500 871 640 839 578 871 b 206 535 321 871 226 703 l 208 535 b 425 631 249 594 339 631 b 735 311 611 631 735 490 b 419 -17 735 118 603 -17 m 199 376 b 418 110 199 251 293 110 b 599 311 532 110 599 197 b 397 514 599 429 524 514 b 199 376 288 514 199 447 "},"7":{"ha":729,"x_min":0,"x_max":0,"o":"m 504 772 b 561 863 535 828 561 863 l 561 865 b 490 861 561 865 539 861 l 194 861 b 164 831 175 861 164 851 l 164 768 l 36 768 l 36 892 b 121 981 36 954 58 981 l 717 981 l 717 886 l 242 0 l 88 0 "},"8":{"ha":800,"x_min":0,"x_max":0,"o":"m 58 281 b 219 532 58 442 219 532 l 219 535 b 99 735 219 535 99 599 b 404 997 99 871 207 997 b 711 732 592 997 711 885 b 593 499 711 588 593 499 l 593 496 b 739 271 682 444 739 376 b 400 -17 739 118 600 -17 b 58 281 215 -17 58 99 m 396 578 b 472 554 415 569 456 554 b 576 724 506 554 576 636 b 404 871 576 814 504 871 b 235 732 297 871 235 810 b 396 578 235 646 308 613 m 199 288 b 400 110 199 178 293 110 b 599 275 503 110 599 178 b 433 436 599 361 525 399 b 332 472 396 451 343 472 b 199 288 297 472 199 392 "},"9":{"ha":788,"x_min":0,"x_max":0,"o":"m 147 143 b 288 110 147 143 210 110 b 582 447 467 110 561 279 l 579 447 b 363 351 539 386 449 351 b 53 669 176 351 53 490 b 368 997 53 863 185 997 b 735 546 576 997 735 819 b 293 -17 735 290 597 -17 b 103 24 176 -17 103 24 m 390 467 b 589 604 500 467 589 533 b 369 871 589 731 494 871 b 189 669 256 871 189 783 b 390 467 189 553 264 467 "}," ":{"ha":374,"x_min":0,"x_max":0,"o":""},"!":{"ha":436,"x_min":0,"x_max":0,"o":"m 149 981 l 288 981 l 279 269 l 158 269 m 150 135 l 288 135 l 288 0 l 150 0 "},"\"":{"ha":472,"x_min":0,"x_max":0,"o":"m 281 997 l 386 997 l 386 736 l 281 736 m 86 997 l 192 997 l 192 736 l 86 736 "},"#":{"ha":981,"x_min":0,"x_max":0,"o":"m 232 265 l 51 265 l 68 368 l 250 368 l 290 601 l 111 601 l 128 703 l 308 703 l 357 981 l 469 981 l 421 703 l 636 703 l 685 981 l 797 981 l 749 703 l 931 703 l 913 601 l 732 601 l 690 368 l 871 368 l 853 265 l 674 265 l 626 0 l 514 0 l 560 265 l 344 265 l 299 0 l 186 0 m 578 368 l 619 601 l 404 601 l 363 368 "},"$":{"ha":758,"x_min":0,"x_max":0,"o":"m 157 256 b 394 110 157 256 238 110 b 563 256 494 110 563 171 b 89 743 563 463 89 426 b 344 993 89 872 193 972 l 344 1124 l 446 1124 l 446 994 b 675 817 544 983 675 935 l 675 736 l 547 736 l 547 782 b 403 871 547 833 481 871 b 231 744 297 871 231 815 b 703 260 231 539 703 590 b 446 -12 703 126 604 7 l 446 -142 l 344 -142 l 344 -14 b 58 174 146 10 58 174 "},"%":{"ha":1088,"x_min":0,"x_max":0,"o":"m 292 586 b 82 792 175 586 82 678 b 292 997 82 904 175 997 b 503 792 408 997 503 904 b 292 586 503 678 408 586 m 853 981 l 983 981 l 233 0 l 103 0 m 292 689 b 394 792 350 689 394 733 b 292 896 394 849 350 896 b 189 792 233 896 189 849 b 292 689 189 733 233 689 m 585 189 b 796 394 585 301 678 394 b 1007 189 911 394 1007 301 b 796 -17 1007 75 911 -17 b 585 189 678 -17 585 75 m 693 189 b 796 86 693 131 738 86 b 899 189 853 86 899 131 b 796 293 899 246 853 293 b 693 189 738 293 693 246 "},"&":{"ha":886,"x_min":0,"x_max":0,"o":"m 61 281 b 231 525 61 382 121 493 l 231 528 b 104 744 231 528 104 574 b 399 997 104 896 222 997 b 522 982 463 997 522 982 l 488 872 b 414 882 488 872 447 882 b 243 731 313 882 243 818 b 424 576 243 669 272 576 l 600 576 l 600 703 l 735 703 l 735 576 l 858 576 l 858 457 l 735 457 l 735 326 b 401 -17 735 108 607 -17 b 61 281 201 -17 61 113 m 201 282 b 401 110 201 189 276 110 b 600 332 522 110 600 178 l 600 457 l 425 457 b 201 282 282 457 201 392 "},"'":{"ha":279,"x_min":0,"x_max":0,"o":"m 86 997 l 193 997 l 193 736 l 86 736 "},"(":{"ha":458,"x_min":0,"x_max":0,"o":"m 115 457 b 272 1021 115 664 167 858 l 390 1021 b 238 467 288 849 238 664 b 397 -132 238 261 292 53 l 278 -132 b 115 457 174 31 115 243 "},")":{"ha":458,"x_min":0,"x_max":0,"o":"m 61 -132 b 221 467 167 53 221 261 b 68 1021 221 664 171 849 l 186 1021 b 343 457 292 858 343 664 b 181 -132 343 243 285 31 l 61 -132 "},"*":{"ha":667,"x_min":0,"x_max":0,"o":"m 124 529 l 247 682 l 247 685 l 58 736 l 94 853 l 279 783 l 269 981 l 396 981 l 385 783 l 571 853 l 608 736 l 418 685 l 418 682 l 540 529 l 442 457 l 335 621 l 332 621 l 224 457 "},"+":{"ha":851,"x_min":0,"x_max":0,"o":"m 58 450 l 369 450 l 369 792 l 483 792 l 483 450 l 793 450 l 793 342 l 483 342 l 483 0 l 369 0 l 369 342 l 58 342 "},",":{"ha":333,"x_min":0,"x_max":0,"o":"m 253 146 l 135 -140 l 31 -140 l 117 146 "},"-":{"ha":599,"x_min":0,"x_max":0,"o":"m 97 456 l 503 456 l 503 336 l 97 336 "},".":{"ha":333,"x_min":0,"x_max":0,"o":"m 97 143 l 235 143 l 235 0 l 97 0 "},"/":{"ha":481,"x_min":0,"x_max":0,"o":"m 354 1033 l 474 1033 l 121 -58 l 1 -58 "},":":{"ha":408,"x_min":0,"x_max":0,"o":"m 136 703 l 274 703 l 274 560 l 136 560 m 136 143 l 274 143 l 274 0 l 136 0 "},";":{"ha":408,"x_min":0,"x_max":0,"o":"m 136 703 l 274 703 l 274 560 l 136 560 m 133 146 l 268 146 l 167 -140 l 63 -140 "},"<":{"ha":851,"x_min":0,"x_max":0,"o":"m 93 442 l 740 731 l 740 607 l 243 397 l 243 394 l 740 185 l 740 61 l 93 350 "},"=":{"ha":851,"x_min":0,"x_max":0,"o":"m 94 581 l 757 581 l 757 472 l 94 472 m 94 318 l 757 318 l 757 210 l 94 210 "},">":{"ha":851,"x_min":0,"x_max":0,"o":"m 103 185 l 599 394 l 599 397 l 103 607 l 103 731 l 750 442 l 750 350 l 103 61 "},"?":{"ha":629,"x_min":0,"x_max":0,"o":"m 207 335 b 443 736 207 546 443 576 b 285 874 443 814 376 874 b 126 815 197 874 126 815 l 51 911 b 294 1000 51 911 143 1000 b 585 746 451 1000 585 900 b 339 326 585 517 339 493 l 339 269 l 207 269 m 204 135 l 342 135 l 342 0 l 204 0 "},"@":{"ha":1058,"x_min":0,"x_max":0,"o":"m 56 335 b 563 849 56 621 279 849 b 924 567 831 849 924 714 l 924 174 l 1024 174 l 1024 72 l 644 72 b 357 335 471 72 357 194 b 644 594 357 472 469 594 l 794 594 b 569 738 792 674 713 738 b 190 335 354 738 190 551 b 585 -64 190 115 349 -64 l 585 -174 b 56 335 274 -174 56 51 m 493 336 b 654 174 493 246 560 174 l 794 174 l 794 501 l 657 501 b 493 336 560 501 493 426 "},"A":{"ha":924,"x_min":0,"x_max":0,"o":"m 17 119 l 36 119 b 82 150 61 119 74 125 l 389 981 l 535 981 l 842 150 b 888 119 850 125 864 119 l 908 119 l 908 0 l 831 0 b 717 81 761 0 739 17 l 643 282 l 282 282 l 208 81 b 93 0 185 17 163 0 l 17 0 m 608 399 l 501 697 b 463 835 483 749 463 835 l 460 835 b 422 697 460 835 440 749 l 315 399 "},"B":{"ha":857,"x_min":0,"x_max":0,"o":"m 142 861 l 51 861 l 51 981 l 471 981 b 756 731 636 981 756 889 b 633 519 756 624 699 551 l 633 517 b 794 279 738 486 794 389 b 479 0 794 94 653 0 l 231 0 b 142 89 168 0 142 26 m 472 568 b 614 717 558 568 614 631 b 469 861 614 803 561 861 l 281 861 l 281 568 m 281 150 b 311 119 281 131 292 119 l 486 119 b 651 286 590 119 651 186 b 486 454 651 386 586 454 l 281 454 l 281 150 "},"C":{"ha":974,"x_min":0,"x_max":0,"o":"m 47 499 b 539 997 47 782 258 997 b 907 785 671 997 907 942 l 907 689 l 779 689 l 779 744 b 542 868 779 826 640 868 b 192 499 344 868 192 719 b 546 114 192 285 346 114 b 856 247 740 114 856 247 l 931 147 b 544 -17 931 147 792 -17 b 47 499 251 -17 47 210 "},"D":{"ha":993,"x_min":0,"x_max":0,"o":"m 143 861 l 53 861 l 53 981 l 450 981 b 946 492 749 981 946 801 b 450 0 946 181 749 0 l 232 0 b 143 89 169 0 143 26 m 282 150 b 313 119 282 131 293 119 l 440 119 b 801 492 658 119 801 250 b 440 861 801 732 657 861 l 282 861 l 282 150 "},"E":{"ha":789,"x_min":0,"x_max":0,"o":"m 143 861 l 53 861 l 53 981 l 629 981 b 718 892 692 981 718 954 l 718 768 l 592 768 l 592 831 b 561 861 592 851 581 861 l 282 861 l 282 554 l 625 554 l 625 435 l 282 435 l 282 150 b 313 119 282 131 293 119 l 599 119 b 629 150 618 119 629 131 l 629 213 l 757 213 l 757 89 b 668 0 757 26 731 0 l 232 0 b 143 89 169 0 143 26 "},"F":{"ha":713,"x_min":0,"x_max":0,"o":"m 143 861 l 53 861 l 53 981 l 593 981 b 682 892 656 981 682 954 l 682 768 l 556 768 l 556 831 b 525 861 556 851 544 861 l 282 861 l 282 539 l 617 539 l 617 419 l 282 419 l 282 0 l 143 0 "},"G":{"ha":1022,"x_min":0,"x_max":0,"o":"m 49 494 b 543 997 49 776 263 997 b 893 876 778 997 893 876 l 814 775 b 546 869 814 775 708 869 b 193 494 350 869 193 713 b 549 114 193 267 350 114 b 828 236 724 114 828 236 l 828 335 b 797 365 828 354 817 365 l 738 365 l 738 485 l 867 485 b 956 396 929 485 956 458 l 956 0 l 833 0 l 833 61 b 835 103 833 82 835 103 l 832 103 b 526 -17 832 103 722 -17 b 49 494 265 -17 49 189 "},"H":{"ha":1063,"x_min":0,"x_max":0,"o":"m 143 831 b 113 861 143 851 132 861 l 53 861 l 53 981 l 193 981 b 282 892 256 981 282 954 l 282 549 l 781 549 l 781 892 b 869 981 781 954 807 981 l 1010 981 l 1010 861 l 950 861 b 919 831 931 861 919 851 l 919 0 l 781 0 l 781 431 l 282 431 l 282 0 l 143 0 "},"I":{"ha":440,"x_min":0,"x_max":0,"o":"m 58 119 l 154 119 l 154 861 l 58 861 l 58 981 l 382 981 l 382 861 l 286 861 l 286 119 l 382 119 l 382 0 l 58 0 "},"J":{"ha":725,"x_min":0,"x_max":0,"o":"m 31 331 l 169 331 l 169 293 b 324 114 169 157 247 114 b 476 288 401 114 476 160 l 476 831 b 446 861 476 851 465 861 l 206 861 l 206 981 l 526 981 b 615 892 589 981 615 954 l 615 282 b 322 -17 615 69 467 -17 b 31 282 178 -17 31 71 "},"K":{"ha":849,"x_min":0,"x_max":0,"o":"m 143 831 b 113 861 143 851 132 861 l 53 861 l 53 981 l 193 981 b 282 892 256 981 282 954 l 282 568 l 375 568 b 456 601 411 568 439 574 l 674 981 l 829 981 l 592 574 b 536 510 564 525 536 510 l 536 507 b 592 446 536 507 564 500 l 739 151 b 808 119 753 122 775 119 l 835 119 l 835 0 l 758 0 b 622 74 682 0 654 10 l 453 414 b 365 450 436 446 406 450 l 282 450 l 282 0 l 143 0 "},"L":{"ha":754,"x_min":0,"x_max":0,"o":"m 143 831 b 113 861 143 851 132 861 l 53 861 l 53 981 l 193 981 b 282 892 256 981 282 954 l 282 150 b 313 119 282 131 293 119 l 583 119 b 614 150 603 119 614 131 l 614 213 l 740 213 l 740 89 b 651 0 740 26 714 0 l 232 0 b 143 89 169 0 143 26 "},"M":{"ha":1190,"x_min":0,"x_max":0,"o":"m 28 119 l 75 119 b 106 150 94 119 104 131 l 172 981 l 315 981 l 546 461 b 594 338 569 408 594 338 l 597 338 b 646 461 597 338 622 408 l 876 981 l 1019 981 l 1086 150 b 1117 119 1088 131 1097 119 l 1163 119 l 1163 0 l 1039 0 b 950 89 976 0 954 26 l 910 607 b 907 751 906 665 907 751 l 904 751 b 853 607 904 751 876 660 l 654 176 l 538 176 l 339 607 b 286 753 315 658 286 753 l 283 753 b 281 607 283 753 285 665 l 242 89 b 151 0 238 26 215 0 l 28 0 "},"N":{"ha":1063,"x_min":0,"x_max":0,"o":"m 53 119 l 113 119 b 143 150 132 119 143 131 l 143 981 l 269 981 l 706 360 b 788 226 742 307 788 226 l 790 226 b 781 360 790 226 781 304 l 781 892 b 871 981 781 954 807 981 l 1010 981 l 1010 861 l 950 861 b 919 831 931 861 919 851 l 919 0 l 794 0 l 357 621 b 276 754 321 674 276 754 l 274 754 b 282 621 274 754 282 675 l 282 89 b 193 0 282 26 256 0 l 53 0 "},"O":{"ha":1101,"x_min":0,"x_max":0,"o":"m 47 497 b 550 997 47 779 267 997 b 1054 497 833 997 1054 779 b 550 -17 1054 208 833 -17 b 47 497 267 -17 47 208 m 192 497 b 550 114 192 281 351 114 b 910 497 749 114 910 281 b 550 868 910 707 749 868 b 192 497 351 868 192 707 "},"P":{"ha":822,"x_min":0,"x_max":0,"o":"m 143 861 l 53 861 l 53 981 l 485 981 b 792 674 664 981 792 860 b 485 363 792 488 664 363 l 282 363 l 282 0 l 143 0 m 463 482 b 649 674 578 482 649 556 b 464 861 649 792 578 861 l 282 861 l 282 482 "},"Q":{"ha":1101,"x_min":0,"x_max":0,"o":"m 47 496 b 549 997 47 779 265 997 b 1054 496 835 997 1054 779 b 939 174 1054 374 1010 260 l 1051 64 l 968 -25 l 856 88 b 549 -17 771 19 664 -17 b 47 496 265 -17 47 210 m 192 496 b 549 114 192 281 349 114 b 756 182 678 114 756 182 l 647 288 l 731 378 l 838 268 b 910 496 882 329 910 411 b 549 867 910 708 751 867 b 192 496 349 867 192 708 "},"R":{"ha":858,"x_min":0,"x_max":0,"o":"m 143 861 l 53 861 l 53 981 l 481 981 b 774 693 653 981 774 872 b 597 433 774 529 667 449 l 597 431 b 639 386 597 431 622 418 l 757 151 b 826 119 771 122 793 119 l 844 119 l 844 0 l 776 0 b 640 74 700 0 672 10 l 507 338 b 410 390 482 383 464 390 l 282 390 l 282 0 l 143 0 m 463 510 b 631 688 568 510 631 578 b 465 861 631 797 568 861 l 282 861 l 282 510 "},"S":{"ha":740,"x_min":0,"x_max":0,"o":"m 129 228 b 375 110 129 228 229 110 b 543 256 467 110 543 165 b 69 735 543 461 69 429 b 383 997 69 882 199 997 b 663 817 488 997 663 950 l 663 736 l 535 736 l 535 782 b 383 871 535 833 461 871 b 211 739 278 871 211 813 b 683 263 211 540 683 585 b 372 -17 683 113 567 -17 b 47 129 164 -17 47 129 "},"T":{"ha":883,"x_min":0,"x_max":0,"o":"m 372 861 l 175 861 b 144 831 156 861 144 851 l 144 768 l 21 768 l 21 897 b 104 981 21 960 42 981 l 779 981 b 863 897 842 981 863 960 l 863 768 l 739 768 l 739 831 b 708 861 739 851 728 861 l 511 861 l 511 0 l 372 0 "},"U":{"ha":1011,"x_min":0,"x_max":0,"o":"m 132 831 b 101 861 132 851 121 861 l 42 861 l 42 981 l 182 981 b 271 892 244 981 271 954 l 271 349 b 506 114 271 201 364 114 b 740 350 647 114 740 201 l 740 892 b 831 981 740 954 768 981 l 969 981 l 969 861 l 910 861 b 879 831 890 861 879 851 l 879 343 b 507 -17 879 129 731 -17 b 132 343 282 -17 132 129 "},"V":{"ha":928,"x_min":0,"x_max":0,"o":"m 82 831 b 36 861 74 856 60 861 l 17 861 l 17 981 l 94 981 b 210 900 164 981 188 965 l 425 296 b 463 161 443 243 463 161 l 465 161 b 503 296 465 161 483 243 l 718 900 b 833 981 742 965 764 981 l 911 981 l 911 861 l 892 861 b 846 831 867 861 854 856 l 538 0 l 390 0 "},"W":{"ha":1344,"x_min":0,"x_max":0,"o":"m 92 831 b 44 861 85 856 69 861 l 24 861 l 24 981 l 103 981 b 224 900 172 981 207 967 l 375 269 b 394 165 388 217 394 165 l 397 165 b 421 269 397 165 406 215 l 618 978 l 743 978 l 926 269 b 947 165 940 215 947 165 l 950 165 b 968 269 950 165 956 217 l 1122 900 b 1242 981 1139 967 1172 981 l 1321 981 l 1321 861 l 1300 861 b 1253 831 1275 861 1260 856 l 1038 0 l 876 0 l 710 619 b 682 754 693 683 682 754 l 679 754 b 647 621 679 754 665 683 l 469 0 l 307 0 "},"X":{"ha":821,"x_min":0,"x_max":0,"o":"m 322 506 l 125 826 b 61 861 106 858 89 861 l 25 861 l 25 981 l 107 981 b 238 906 174 981 201 968 l 347 719 b 410 604 378 665 410 604 l 413 604 b 472 719 413 604 440 665 l 583 906 b 715 981 619 968 647 981 l 796 981 l 796 861 l 761 861 b 696 826 732 861 715 858 l 497 506 l 807 0 l 650 0 l 472 299 b 410 410 440 353 410 410 l 407 410 b 347 297 407 410 379 351 l 171 0 l 14 0 "},"Y":{"ha":838,"x_min":0,"x_max":0,"o":"m 350 428 l 101 826 b 42 861 83 856 71 861 l 14 861 l 14 981 l 78 981 b 210 906 144 981 171 971 l 358 658 b 418 549 390 604 418 549 l 421 549 b 481 658 421 549 447 604 l 628 906 b 760 981 667 971 694 981 l 824 981 l 824 861 l 797 861 b 736 826 768 861 754 856 l 489 428 l 489 0 l 350 0 "},"Z":{"ha":808,"x_min":0,"x_max":0,"o":"m 25 92 l 508 772 b 582 863 547 826 582 863 l 582 865 b 508 861 582 865 556 861 l 193 861 b 163 831 174 861 163 851 l 163 768 l 36 768 l 36 892 b 125 981 36 954 63 981 l 754 981 l 754 890 l 269 208 b 197 119 232 154 197 119 l 197 117 b 269 119 197 117 224 119 l 615 119 b 646 150 635 119 646 131 l 646 213 l 774 213 l 774 89 b 685 0 774 26 747 0 l 25 0 "},"[":{"ha":454,"x_min":0,"x_max":0,"o":"m 135 932 b 224 1021 135 994 161 1021 l 364 1021 l 364 919 l 283 919 b 253 889 264 919 253 910 l 253 1 b 283 -29 253 -19 264 -29 l 364 -29 l 364 -132 l 224 -132 b 135 -43 161 -132 135 -106 "},"\\":{"ha":472,"x_min":0,"x_max":0,"o":"m 7 1033 l 126 1033 l 479 -58 l 360 -58 "},"]":{"ha":454,"x_min":0,"x_max":0,"o":"m 171 -29 b 201 1 190 -29 201 -19 l 201 889 b 171 919 201 910 190 919 l 90 919 l 90 1021 l 231 1021 b 321 932 294 1021 321 994 l 321 -43 b 231 -132 321 -106 294 -132 l 90 -132 l 90 -29 "},"^":{"ha":851,"x_min":0,"x_max":0,"o":"m 374 981 l 461 981 l 738 343 l 617 343 l 417 832 l 218 343 l 96 343 "},"_":{"ha":761,"x_min":0,"x_max":0,"o":"m 754 0 l 754 -108 l 7 -108 l 7 0 "},"`":{"ha":526,"x_min":0,"x_max":0,"o":"m 115 1225 l 251 1225 l 353 1049 l 247 1049 "},"a":{"ha":756,"x_min":0,"x_max":0,"o":"m 51 197 b 475 438 51 424 357 438 l 506 438 l 506 444 b 350 613 506 568 460 613 b 235 558 317 613 235 604 l 235 518 l 108 518 l 108 588 b 351 719 108 703 283 719 b 640 444 579 719 640 600 l 640 146 b 671 115 640 125 651 115 l 728 115 l 728 0 l 601 0 b 514 81 539 0 514 31 b 517 122 514 104 517 122 l 514 122 b 290 -17 514 122 458 -17 b 51 197 169 -17 51 57 m 188 204 b 319 92 188 147 231 92 b 507 315 433 92 507 211 l 507 339 l 472 339 b 188 204 371 339 188 333 "},"b":{"ha":833,"x_min":0,"x_max":0,"o":"m 113 835 b 82 865 113 856 101 865 l 25 865 l 25 981 l 158 981 b 247 892 221 981 247 954 l 247 663 b 244 604 247 629 244 604 l 247 604 b 474 719 247 604 308 719 b 782 351 663 719 782 569 b 463 -17 782 128 647 -17 b 242 101 306 -17 242 101 l 239 101 b 242 49 239 101 242 79 l 242 0 l 113 0 m 244 349 b 443 99 244 224 310 99 b 646 350 554 99 646 190 b 446 601 646 503 564 601 b 244 349 342 601 244 528 "},"c":{"ha":740,"x_min":0,"x_max":0,"o":"m 50 351 b 428 719 50 558 201 719 b 681 565 513 719 681 685 l 681 493 l 558 493 l 558 532 b 428 604 558 582 479 604 b 188 353 290 604 188 501 b 435 103 188 193 304 103 b 653 200 567 103 653 200 l 711 104 b 425 -17 711 104 611 -17 b 50 351 204 -17 50 140 "},"d":{"ha":840,"x_min":0,"x_max":0,"o":"m 51 351 b 369 719 51 575 183 719 b 586 611 532 719 586 611 l 589 611 b 586 658 589 611 586 632 l 586 835 b 556 865 586 856 575 865 l 499 865 l 499 981 l 632 981 b 721 892 694 981 721 954 l 721 146 b 751 115 721 125 732 115 l 808 115 l 808 0 l 679 0 b 593 75 618 0 593 28 b 594 111 593 96 594 111 l 592 111 b 361 -17 592 111 533 -17 b 51 351 171 -17 51 133 m 388 100 b 590 353 492 100 590 174 b 390 603 590 478 525 603 b 189 351 281 603 189 511 b 388 100 189 199 271 100 "},"e":{"ha":767,"x_min":0,"x_max":0,"o":"m 50 351 b 403 719 50 575 203 719 b 706 390 600 719 706 572 b 701 335 706 372 701 335 l 189 335 b 431 100 197 185 306 100 b 636 182 551 100 636 182 l 694 86 b 424 -17 694 86 590 -17 b 50 351 201 -17 50 143 m 567 433 b 401 613 563 550 489 613 b 194 433 299 613 215 550 "},"f":{"ha":451,"x_min":0,"x_max":0,"o":"m 118 594 l 31 594 l 31 703 l 118 703 l 118 726 b 383 988 118 958 294 988 b 435 983 414 988 435 983 l 435 868 b 400 871 435 868 421 871 b 253 724 347 871 253 854 l 253 703 l 419 703 l 419 594 l 253 594 l 253 0 l 118 0 "},"g":{"ha":826,"x_min":0,"x_max":0,"o":"m 51 368 b 360 719 51 569 164 719 b 585 610 535 719 585 610 l 588 610 b 588 632 588 610 588 621 b 665 703 588 674 610 703 l 796 703 l 796 588 l 740 588 b 710 560 722 588 710 578 l 710 29 b 356 -292 710 -203 533 -292 b 115 -233 271 -292 185 -268 l 161 -125 b 353 -172 161 -125 247 -172 b 575 24 475 -172 575 -118 l 575 71 b 578 115 575 93 578 115 l 575 115 b 365 8 533 47 468 8 b 51 368 171 8 51 165 m 392 125 b 579 367 489 125 579 183 b 381 603 579 550 489 603 b 189 371 258 603 189 514 b 392 125 189 225 265 125 "},"h":{"ha":864,"x_min":0,"x_max":0,"o":"m 113 835 b 82 865 113 856 101 865 l 25 865 l 25 981 l 158 981 b 247 894 221 981 247 954 l 247 621 b 244 563 247 588 244 563 l 247 563 b 508 719 278 631 368 719 b 749 450 674 719 749 629 l 749 146 b 779 115 749 125 760 115 l 836 115 l 836 0 l 704 0 b 615 89 642 0 615 26 l 615 419 b 482 596 615 518 594 596 b 260 422 375 596 289 524 b 247 329 250 394 247 363 l 247 0 l 113 0 "},"i":{"ha":382,"x_min":0,"x_max":0,"o":"m 128 981 l 249 981 l 249 840 l 128 840 m 125 557 b 94 588 125 578 114 588 l 38 588 l 38 703 l 169 703 b 258 614 232 703 258 676 l 258 146 b 289 115 258 125 269 115 l 346 115 l 346 0 l 214 0 b 125 89 151 0 125 26 "},"j":{"ha":397,"x_min":0,"x_max":0,"o":"m 156 981 l 278 981 l 278 840 l 156 840 m -28 -168 b 3 -169 -28 -168 -15 -169 b 153 -15 54 -169 153 -153 l 153 557 b 122 588 153 578 142 588 l 65 588 l 65 703 l 199 703 b 288 614 261 703 288 676 l 288 -24 b 18 -286 288 -258 107 -286 b -28 -283 -10 -286 -28 -283 l -28 -168 "},"k":{"ha":729,"x_min":0,"x_max":0,"o":"m 113 835 b 82 865 113 856 101 865 l 25 865 l 25 981 l 158 981 b 247 900 221 981 247 963 l 247 444 l 310 444 b 383 475 335 444 365 451 l 546 703 l 700 703 l 510 442 b 457 389 478 399 457 389 l 457 386 b 504 333 457 386 481 378 l 608 143 b 668 115 619 121 633 115 l 711 115 l 711 0 l 614 0 b 497 65 549 0 528 11 l 369 301 b 300 331 354 328 322 331 l 247 331 l 247 0 l 113 0 "},"l":{"ha":372,"x_min":0,"x_max":0,"o":"m 115 835 b 85 865 115 856 104 865 l 28 865 l 28 981 l 161 981 b 250 892 224 981 250 954 l 250 146 b 281 115 250 125 261 115 l 338 115 l 338 0 l 204 0 b 115 89 142 0 115 26 "},"m":{"ha":1301,"x_min":0,"x_max":0,"o":"m 118 557 b 88 588 118 578 107 588 l 31 588 l 31 703 l 160 703 b 249 625 221 703 249 676 l 249 599 b 247 558 249 576 247 558 l 250 558 b 490 719 283 640 381 719 b 701 560 610 719 678 665 l 704 560 b 954 719 743 644 840 719 b 1186 450 1113 719 1186 629 l 1186 146 b 1217 115 1186 125 1197 115 l 1274 115 l 1274 0 l 1140 0 b 1051 89 1078 0 1051 26 l 1051 421 b 928 597 1051 519 1032 597 b 731 417 831 597 756 513 b 719 321 724 388 719 357 l 719 0 l 585 0 l 585 421 b 464 597 585 513 571 597 b 263 411 361 597 289 511 b 253 321 256 383 253 353 l 253 0 l 118 0 "},"n":{"ha":869,"x_min":0,"x_max":0,"o":"m 118 557 b 88 588 118 578 107 588 l 31 588 l 31 703 l 160 703 b 249 625 219 703 249 676 l 249 599 b 246 558 249 576 246 558 l 249 558 b 513 719 276 619 360 719 b 754 450 679 719 754 629 l 754 146 b 785 115 754 125 765 115 l 842 115 l 842 0 l 708 0 b 619 89 646 0 619 26 l 619 419 b 488 596 619 518 600 596 b 265 424 379 596 293 525 b 253 329 256 396 253 364 l 253 0 l 118 0 "},"o":{"ha":857,"x_min":0,"x_max":0,"o":"m 50 353 b 428 719 50 565 218 719 b 807 353 638 719 807 565 b 428 -17 807 139 638 -17 b 50 353 218 -17 50 139 m 188 353 b 428 100 188 206 296 100 b 669 353 561 100 669 206 b 428 603 669 499 561 603 b 188 353 296 603 188 499 "},"p":{"ha":839,"x_min":0,"x_max":0,"o":"m 118 557 b 88 588 118 578 107 588 l 31 588 l 31 703 l 157 703 b 243 632 219 703 243 676 b 242 594 243 611 242 594 l 244 594 b 478 719 244 594 303 719 b 788 351 667 719 788 569 b 468 -17 788 128 651 -17 b 253 96 313 -17 253 96 l 250 96 b 253 35 250 96 253 71 l 253 -278 l 118 -278 m 249 349 b 447 99 249 224 317 99 b 650 350 558 99 650 190 b 451 601 650 503 568 601 b 249 349 347 601 249 528 "},"q":{"ha":839,"x_min":0,"x_max":0,"o":"m 53 351 b 369 719 53 575 185 719 b 596 600 536 719 596 600 l 599 600 b 597 636 599 600 597 615 b 683 703 597 678 624 703 l 810 703 l 810 588 l 753 588 b 722 557 733 588 722 578 l 722 -278 l 588 -278 l 588 40 b 590 101 588 76 590 101 l 588 101 b 361 -17 588 101 528 -17 b 53 351 172 -17 53 133 m 389 100 b 590 353 493 100 590 174 b 392 603 590 478 525 603 b 189 351 281 603 189 511 b 389 100 189 199 271 100 "},"r":{"ha":538,"x_min":0,"x_max":0,"o":"m 118 557 b 88 588 118 578 107 588 l 31 588 l 31 703 l 158 703 b 249 621 219 703 249 678 l 249 576 b 246 528 249 549 246 528 l 249 528 b 471 711 282 631 363 711 b 507 708 489 711 507 708 l 507 575 b 467 579 507 575 489 579 b 269 414 381 579 301 518 b 253 288 257 374 253 331 l 253 0 l 118 0 "},"s":{"ha":631,"x_min":0,"x_max":0,"o":"m 107 197 b 319 93 107 197 190 93 b 442 183 385 93 442 122 b 58 522 442 310 58 296 b 319 719 58 657 176 719 b 550 583 404 719 550 690 l 550 517 l 429 517 l 429 551 b 324 611 429 593 367 611 b 193 528 244 611 193 583 b 576 189 193 394 576 424 b 318 -17 576 64 465 -17 b 38 107 131 -17 38 107 "},"t":{"ha":489,"x_min":0,"x_max":0,"o":"m 124 594 l 32 594 l 32 703 l 126 703 l 126 896 l 258 896 l 258 703 l 426 703 l 426 594 l 258 594 l 258 272 b 407 114 258 132 356 114 b 439 117 426 114 439 117 l 439 -3 b 392 -6 439 -3 421 -6 b 124 257 303 -6 124 22 "},"u":{"ha":863,"x_min":0,"x_max":0,"o":"m 115 557 b 85 588 115 578 104 588 l 28 588 l 28 703 l 160 703 b 250 617 224 703 250 676 l 250 283 b 381 108 250 186 269 108 b 610 374 525 108 610 235 l 610 703 l 744 703 l 744 146 b 775 115 744 125 756 115 l 832 115 l 832 0 l 703 0 b 614 78 643 0 614 28 l 614 104 b 615 144 614 126 615 144 l 613 144 b 354 -17 583 75 493 -17 b 115 253 196 -17 115 67 "},"v":{"ha":739,"x_min":0,"x_max":0,"o":"m 75 560 b 32 588 67 581 54 588 l 18 588 l 18 703 l 88 703 b 194 632 150 703 174 688 l 346 228 b 371 136 360 189 371 136 l 374 136 b 399 228 374 136 385 189 l 549 632 b 656 703 569 688 594 703 l 721 703 l 721 588 l 707 588 b 661 560 683 588 669 581 l 447 0 l 292 0 "},"w":{"ha":1158,"x_min":0,"x_max":0,"o":"m 76 560 b 33 588 69 582 56 588 l 14 588 l 14 703 l 99 703 b 203 628 161 703 185 688 l 328 210 b 344 135 339 174 344 135 l 347 135 b 367 210 347 135 356 174 l 518 700 l 642 700 l 794 210 b 813 135 806 174 813 135 l 815 135 b 832 210 815 135 821 174 l 956 628 b 1061 703 974 688 999 703 l 1144 703 l 1144 588 l 1125 588 b 1082 560 1103 588 1089 582 l 897 0 l 746 0 l 600 451 b 581 528 589 488 581 528 l 578 528 b 558 451 578 528 569 488 l 414 0 l 261 0 "},"x":{"ha":686,"x_min":0,"x_max":0,"o":"m 258 361 l 124 561 b 64 588 108 583 93 588 l 25 588 l 25 703 l 111 703 b 233 632 176 703 196 690 l 319 496 b 342 451 329 478 342 451 l 344 451 b 368 496 344 451 357 478 l 453 632 b 575 703 489 692 510 703 l 663 703 l 663 588 l 622 588 b 563 563 594 588 576 583 l 428 361 l 665 0 l 514 0 l 365 229 b 344 267 356 246 344 267 l 342 267 b 319 229 342 267 331 246 l 174 0 l 21 0 "},"y":{"ha":758,"x_min":0,"x_max":0,"o":"m 63 -125 b 165 -175 63 -126 106 -175 b 289 -74 219 -175 263 -136 l 321 3 l 81 560 b 36 588 71 582 58 588 l 19 588 l 19 703 l 99 703 b 199 632 156 703 176 688 l 358 236 b 383 151 372 200 383 151 l 386 151 b 410 236 386 151 396 200 l 556 632 b 658 703 576 688 601 703 l 739 703 l 739 588 l 721 588 b 676 560 699 588 686 582 l 399 -126 b 172 -292 356 -235 271 -292 b 7 -222 68 -292 7 -222 l 63 -125 "},"z":{"ha":701,"x_min":0,"x_max":0,"o":"m 46 78 l 406 522 b 464 588 436 560 464 588 l 464 590 b 396 588 464 590 442 588 l 204 588 b 174 557 185 588 174 576 l 174 510 l 50 510 l 50 614 b 139 703 50 676 76 703 l 642 703 l 642 625 l 281 179 b 222 115 251 144 222 115 l 222 113 b 290 115 222 113 244 115 l 515 115 b 546 146 535 115 546 126 l 546 193 l 668 193 l 668 89 b 579 0 668 26 642 0 l 46 0 "},"{":{"ha":471,"x_min":0,"x_max":0,"o":"m 158 240 b 40 392 158 382 40 392 l 40 511 b 158 665 40 511 158 521 l 158 782 b 383 1024 158 1001 317 1024 b 415 1022 403 1024 415 1022 l 415 919 b 396 919 415 919 407 919 b 276 774 361 919 276 907 l 276 631 b 164 454 276 493 164 454 l 164 451 b 276 275 164 451 276 418 l 276 115 b 396 -31 276 -18 361 -31 b 415 -31 407 -31 415 -31 l 415 -133 b 383 -136 415 -133 403 -136 b 158 107 317 -136 158 -114 "},"|":{"ha":415,"x_min":0,"x_max":0,"o":"m 149 1100 l 267 1100 l 267 -221 l 149 -221 "},"}":{"ha":471,"x_min":0,"x_max":0,"o":"m 56 -31 b 75 -31 56 -31 63 -31 b 194 115 110 -31 194 -18 l 194 275 b 307 451 194 418 307 451 l 307 454 b 194 631 307 454 194 493 l 194 774 b 75 919 194 907 110 919 b 56 919 63 919 56 919 l 56 1022 b 88 1024 56 1022 68 1024 b 313 782 154 1024 313 1001 l 313 665 b 431 511 313 521 431 511 l 431 392 b 313 240 431 392 313 382 l 313 107 b 88 -136 313 -114 154 -136 b 56 -133 68 -136 56 -133 l 56 -31 "},"~":{"ha":851,"x_min":0,"x_max":0,"o":"m 86 283 b 281 518 86 444 160 518 b 546 382 426 518 442 382 b 639 513 619 382 639 454 l 749 513 b 556 278 749 350 675 278 b 289 413 410 278 394 413 b 196 283 217 413 196 340 l 86 283 "},"¡":{"ha":436,"x_min":0,"x_max":0,"o":"m 149 703 l 286 703 l 286 568 l 149 568 m 158 433 l 279 433 l 288 -278 l 149 -278 "},"¢":{"ha":785,"x_min":0,"x_max":0,"o":"m 360 101 b 76 490 182 126 76 290 b 360 881 76 692 182 856 l 360 1033 l 461 1033 l 461 881 b 713 665 581 863 667 781 l 590 618 b 415 756 558 696 501 756 b 214 490 286 756 214 640 b 415 226 214 336 283 226 b 590 364 501 226 558 285 l 713 315 b 461 101 671 213 588 115 l 461 -53 l 360 -53 "},"£":{"ha":811,"x_min":0,"x_max":0,"o":"m 69 119 l 154 119 l 154 443 l 93 443 l 93 540 l 154 540 l 154 729 b 456 997 154 883 281 997 b 688 908 596 997 688 908 l 608 813 b 457 872 608 813 542 872 b 293 724 354 872 293 806 l 293 540 l 549 540 l 549 443 l 293 443 l 293 119 l 593 119 b 624 150 613 119 624 131 l 624 213 l 751 213 l 751 89 b 661 0 751 26 725 0 l 69 0 "},"¥":{"ha":853,"x_min":0,"x_max":0,"o":"m 133 371 l 357 371 l 357 428 l 318 489 l 131 489 l 131 576 l 261 576 l 261 579 l 106 831 b 43 861 89 857 72 861 l 21 861 l 21 981 l 82 981 b 217 906 149 981 178 969 l 365 658 b 425 550 397 604 425 550 l 428 550 b 488 658 428 550 456 604 l 635 906 b 771 981 674 971 704 981 l 832 981 l 832 861 l 808 861 b 746 831 779 861 763 857 l 592 579 l 592 576 l 718 576 l 718 489 l 533 489 l 496 428 l 496 371 l 718 371 l 718 283 l 496 283 l 496 0 l 357 0 l 357 283 l 133 283 "},"§":{"ha":638,"x_min":0,"x_max":0,"o":"m 114 696 b 106 760 108 717 106 739 b 368 997 106 894 204 997 b 564 926 488 997 564 926 l 500 838 b 368 885 500 838 444 885 b 219 744 285 885 219 839 b 228 689 219 728 222 710 l 322 154 l 211 154 m 146 78 b 281 26 146 78 203 26 b 428 172 363 26 428 75 b 419 235 428 190 425 211 l 328 757 l 438 757 l 535 215 b 542 153 539 193 542 172 b 281 -86 542 15 443 -86 b 85 -15 160 -86 85 -15 "},"¨":{"ha":696,"x_min":0,"x_max":0,"o":"m 426 1225 l 528 1225 l 528 1085 l 426 1085 m 197 1225 l 300 1225 l 300 1085 l 197 1085 "},"©":{"ha":1093,"x_min":0,"x_max":0,"o":"m 51 490 b 544 997 51 772 271 997 b 1042 490 821 997 1042 772 b 544 -17 1042 208 821 -17 b 51 490 271 -17 51 208 m 153 490 b 544 78 153 258 324 78 b 940 490 768 78 940 258 b 544 903 940 722 768 903 b 153 490 324 903 153 722 m 281 489 b 554 767 281 635 386 767 b 781 629 714 767 781 629 l 696 583 b 558 671 696 583 649 671 b 386 489 449 671 386 585 b 558 307 386 390 446 307 b 696 394 650 307 696 394 l 781 349 b 554 211 781 349 714 211 b 281 489 386 211 281 342 "},"ª":{"ha":576,"x_min":0,"x_max":0,"o":"m 228 532 b 76 669 150 532 76 579 b 332 818 76 807 253 818 l 351 818 b 267 910 349 888 326 910 b 208 886 251 910 208 908 l 208 860 l 111 860 l 111 915 b 268 994 111 983 225 994 b 451 819 414 994 451 917 l 451 644 b 465 629 451 635 456 629 l 504 629 l 504 540 l 424 540 b 360 592 389 540 364 553 l 358 606 l 356 606 b 228 532 356 606 319 532 m 249 614 b 351 747 313 614 349 678 l 338 747 b 179 674 282 747 179 743 b 249 614 179 643 201 614 m 75 447 l 508 447 l 508 368 l 75 368 "},"«":{"ha":788,"x_min":0,"x_max":0,"o":"m 565 678 l 700 678 l 472 392 l 700 107 l 565 107 l 338 392 m 58 392 l 286 678 l 421 678 l 193 392 l 421 107 l 286 107 "},"¬":{"ha":851,"x_min":0,"x_max":0,"o":"m 94 581 l 757 581 l 757 210 l 644 210 l 644 472 l 94 472 "},"®":{"ha":1093,"x_min":0,"x_max":0,"o":"m 51 490 b 544 997 51 772 271 997 b 1042 490 821 997 1042 772 b 544 -17 1042 208 821 -17 b 51 490 271 -17 51 208 m 153 490 b 544 78 153 258 324 78 b 940 490 768 78 940 258 b 544 903 940 722 768 903 b 153 490 324 903 153 722 m 375 679 l 328 679 l 328 754 l 576 754 b 742 594 675 754 742 692 b 646 456 742 514 689 467 l 646 453 b 675 425 646 453 663 449 l 729 321 b 763 303 739 304 751 303 l 778 303 l 778 231 l 725 231 b 650 269 686 231 668 235 l 578 407 b 531 433 565 431 551 433 l 474 433 l 474 231 l 375 231 m 557 506 b 642 594 608 506 642 539 b 557 679 642 649 608 679 l 474 679 l 474 506 "},"¯":{"ha":644,"x_min":0,"x_max":0,"o":"m 150 1183 l 503 1183 l 503 1089 l 150 1089 "},"°":{"ha":551,"x_min":0,"x_max":0,"o":"m 275 586 b 65 792 158 586 65 678 b 275 997 65 904 158 997 b 486 792 392 997 486 904 b 275 586 486 678 392 586 m 275 689 b 378 792 333 689 378 735 b 275 896 378 849 333 896 b 172 792 217 896 172 849 b 275 689 172 735 217 689 "},"±":{"ha":851,"x_min":0,"x_max":0,"o":"m 58 450 l 369 450 l 369 792 l 483 792 l 483 450 l 793 450 l 793 342 l 483 342 l 483 0 l 369 0 l 369 342 l 58 342 m 771 -132 l 771 -240 l 81 -240 l 81 -132 "},"²":{"ha":588,"x_min":0,"x_max":0,"o":"m 138 369 b 71 451 89 369 71 385 b 243 671 71 571 161 625 b 371 813 310 708 371 744 b 279 894 371 860 336 894 b 199 846 232 894 199 871 l 199 813 l 100 813 l 100 874 b 281 993 100 958 215 993 b 478 813 396 993 478 921 b 297 594 478 693 381 640 b 179 471 236 560 179 526 b 192 463 179 464 185 463 l 399 463 b 407 472 404 463 407 465 l 407 518 l 504 518 l 504 433 b 440 369 504 389 485 369 l 138 369 "},"³":{"ha":588,"x_min":0,"x_max":0,"o":"m 289 358 b 72 469 158 358 72 469 l 146 538 b 289 457 146 538 206 457 b 394 557 346 457 394 499 b 282 650 394 615 344 650 l 235 650 l 218 711 l 335 858 b 365 890 347 875 365 890 l 365 893 b 318 890 365 893 342 890 l 213 890 b 200 876 204 890 200 886 l 200 840 l 101 840 l 101 914 b 168 981 101 964 117 981 l 490 981 l 490 910 l 346 733 b 501 556 426 722 501 653 b 289 358 501 438 401 358 "},"´":{"ha":539,"x_min":0,"x_max":0,"o":"m 267 1225 l 403 1225 l 271 1049 l 165 1049 "},"µ":{"ha":863,"x_min":0,"x_max":0,"o":"m 28 703 l 160 703 b 250 617 224 703 250 676 l 250 283 b 381 108 250 186 269 108 b 610 374 525 108 610 235 l 610 703 l 744 703 l 744 146 b 775 115 744 125 756 115 l 832 115 l 832 0 l 703 0 b 614 78 642 0 614 28 l 614 104 b 615 144 614 126 615 144 l 613 144 b 354 -17 613 144 547 -17 b 228 26 261 -17 228 26 l 225 26 b 235 -82 225 26 235 -42 l 235 -278 l 115 -278 l 115 557 b 85 588 115 578 104 588 l 28 588 "},"¶":{"ha":832,"x_min":0,"x_max":0,"o":"m 386 335 b 57 658 207 335 57 476 b 385 981 57 840 206 981 l 763 981 l 763 861 l 490 861 l 490 -69 l 386 -69 m 576 782 l 679 782 l 679 -69 l 576 -69 "},"·":{"ha":333,"x_min":0,"x_max":0,"o":"m 97 475 l 235 475 l 235 332 l 97 332 "},"¸":{"ha":578,"x_min":0,"x_max":0,"o":"m 290 18 l 290 7 l 275 -69 b 374 -171 331 -76 374 -111 b 232 -287 374 -257 304 -287 b 156 -278 189 -287 156 -278 l 156 -197 b 215 -207 156 -197 183 -207 b 281 -164 250 -207 281 -196 b 207 -121 281 -135 250 -121 b 183 -122 196 -121 183 -122 l 222 32 "},"¹":{"ha":588,"x_min":0,"x_max":0,"o":"m 106 463 l 235 463 l 235 797 b 236 843 235 826 236 843 l 233 843 b 217 819 233 843 228 829 l 161 767 l 97 835 l 249 981 l 340 981 l 340 463 l 468 463 l 468 369 l 106 369 "},"º":{"ha":633,"x_min":0,"x_max":0,"o":"m 317 533 b 79 767 185 533 79 631 b 315 996 79 899 185 996 b 554 767 447 996 554 899 b 317 533 554 631 447 533 m 317 622 b 453 767 392 622 453 682 b 317 907 453 849 392 907 b 179 767 240 907 179 849 b 317 622 179 682 240 622 m 101 447 l 535 447 l 535 368 l 101 368 "},"»":{"ha":788,"x_min":0,"x_max":0,"o":"m 594 392 l 367 678 l 501 678 l 729 392 l 501 107 l 367 107 m 314 392 l 86 678 l 221 678 l 449 392 l 221 107 l 86 107 "},"¼":{"ha":1272,"x_min":0,"x_max":0,"o":"m 793 981 l 904 981 l 507 0 l 394 0 m 108 463 l 238 463 l 238 797 b 239 843 238 826 239 843 l 236 843 b 219 819 236 843 229 829 l 164 767 l 100 835 l 250 981 l 343 981 l 343 463 l 467 463 l 467 369 l 108 369 m 718 221 l 983 613 l 1104 613 l 1104 250 l 1196 250 l 1196 157 l 1104 157 l 1104 0 l 999 0 l 999 157 l 718 157 m 832 249 b 860 250 832 249 850 250 l 1000 250 l 1000 440 b 1001 490 1000 465 1001 490 l 999 490 b 978 453 999 490 990 471 l 851 274 b 832 251 843 263 832 251 l 832 249 "},"½":{"ha":1272,"x_min":0,"x_max":0,"o":"m 778 981 l 889 981 l 490 0 l 379 0 m 108 463 l 238 463 l 238 797 b 239 843 238 826 239 843 l 236 843 b 219 819 236 843 229 829 l 164 767 l 100 835 l 250 981 l 343 981 l 343 463 l 467 463 l 467 369 l 108 369 m 754 82 b 926 303 754 201 844 256 b 1054 443 992 339 1054 375 b 961 526 1054 492 1018 526 b 882 478 915 526 882 503 l 882 444 l 783 444 l 783 504 b 964 624 783 589 899 624 b 1161 444 1079 624 1161 551 b 981 225 1161 324 1063 271 b 861 101 918 190 861 157 b 875 93 861 94 867 93 l 1081 93 b 1089 103 1086 93 1089 96 l 1089 150 l 1188 150 l 1188 65 b 1122 0 1188 19 1168 0 l 819 0 b 754 82 771 0 754 17 "},"¾":{"ha":1272,"x_min":0,"x_max":0,"o":"m 790 981 l 901 981 l 504 0 l 392 0 m 300 358 b 83 469 169 358 83 469 l 156 538 b 300 457 156 538 217 457 b 406 557 357 457 406 499 b 293 650 406 615 356 650 l 244 650 l 228 711 l 346 858 b 375 890 358 875 375 890 l 375 893 b 329 890 375 893 353 890 l 224 890 b 211 876 214 890 211 886 l 211 840 l 113 840 l 113 914 b 178 981 113 964 126 981 l 501 981 l 501 910 l 356 733 b 513 556 438 722 513 653 b 300 358 513 438 413 358 m 718 221 l 983 613 l 1104 613 l 1104 250 l 1196 250 l 1196 157 l 1104 157 l 1104 0 l 999 0 l 999 157 l 718 157 m 832 249 b 860 250 832 249 850 250 l 1000 250 l 1000 440 b 1001 490 1000 465 1001 490 l 999 490 b 978 453 999 490 990 471 l 851 274 b 832 251 843 263 832 251 l 832 249 "},"¿":{"ha":629,"x_min":0,"x_max":0,"o":"m 288 703 l 425 703 l 425 568 l 288 568 m 44 -39 b 290 382 44 190 290 214 l 290 433 l 421 433 l 421 374 b 186 -28 421 163 186 131 b 344 -167 186 -106 253 -167 b 503 -107 432 -167 503 -107 l 576 -203 b 335 -292 576 -203 486 -292 b 44 -39 178 -292 44 -193 "},"À":{"ha":924,"x_min":0,"x_max":0,"o":"m 17 119 l 36 119 b 82 150 61 119 74 125 l 389 981 l 535 981 l 842 150 b 888 119 850 125 864 119 l 908 119 l 908 0 l 831 0 b 717 81 761 0 739 17 l 643 282 l 282 282 l 208 81 b 93 0 185 17 163 0 l 17 0 m 608 399 l 501 697 b 463 835 483 749 463 835 l 460 835 b 422 697 460 835 440 749 l 315 399 m 276 1225 l 413 1225 l 514 1049 l 408 1049 "},"Á":{"ha":924,"x_min":0,"x_max":0,"o":"m 17 119 l 36 119 b 82 150 61 119 74 125 l 389 981 l 535 981 l 842 150 b 888 119 850 125 864 119 l 908 119 l 908 0 l 831 0 b 717 81 761 0 739 17 l 643 282 l 282 282 l 208 81 b 93 0 185 17 163 0 l 17 0 m 608 399 l 501 697 b 463 835 483 749 463 835 l 460 835 b 422 697 460 835 440 749 l 315 399 m 510 1225 l 646 1225 l 514 1049 l 408 1049 "},"Â":{"ha":924,"x_min":0,"x_max":0,"o":"m 17 119 l 36 119 b 82 150 61 119 74 125 l 389 981 l 535 981 l 842 150 b 888 119 850 125 864 119 l 908 119 l 908 0 l 831 0 b 717 81 761 0 739 17 l 643 282 l 282 282 l 208 81 b 93 0 185 17 163 0 l 17 0 m 608 399 l 501 697 b 463 835 483 749 463 835 l 460 835 b 422 697 460 835 440 749 l 315 399 m 400 1225 l 524 1225 l 646 1049 l 536 1049 l 464 1160 l 461 1160 l 388 1049 l 279 1049 "},"Ã":{"ha":924,"x_min":0,"x_max":0,"o":"m 17 119 l 36 119 b 82 150 61 119 74 125 l 389 981 l 535 981 l 842 150 b 888 119 850 125 864 119 l 908 119 l 908 0 l 831 0 b 717 81 761 0 739 17 l 643 282 l 282 282 l 208 81 b 93 0 185 17 163 0 l 17 0 m 608 399 l 501 697 b 463 835 483 749 463 835 l 460 835 b 422 697 460 835 440 749 l 315 399 m 226 1050 b 369 1226 226 1167 275 1226 b 547 1132 465 1226 486 1132 b 603 1222 589 1132 603 1182 l 694 1222 b 551 1046 694 1107 646 1046 b 374 1140 457 1046 435 1140 b 318 1050 332 1140 318 1092 l 226 1050 "},"Ä":{"ha":924,"x_min":0,"x_max":0,"o":"m 17 119 l 36 119 b 82 150 61 119 74 125 l 389 981 l 535 981 l 842 150 b 888 119 850 125 864 119 l 908 119 l 908 0 l 831 0 b 717 81 761 0 739 17 l 643 282 l 282 282 l 208 81 b 93 0 185 17 163 0 l 17 0 m 608 399 l 501 697 b 463 835 483 749 463 835 l 460 835 b 422 697 460 835 440 749 l 315 399 m 526 1225 l 628 1225 l 628 1085 l 526 1085 m 297 1225 l 400 1225 l 400 1085 l 297 1085 "},"Å":{"ha":924,"x_min":0,"x_max":0,"o":"m 17 119 l 36 119 b 82 150 61 119 74 125 l 389 981 l 535 981 l 842 150 b 888 119 850 125 864 119 l 908 119 l 908 0 l 831 0 b 717 81 761 0 739 17 l 643 282 l 282 282 l 208 81 b 93 0 185 17 163 0 l 17 0 m 608 399 l 501 697 b 463 835 483 749 463 835 l 460 835 b 422 697 460 835 440 749 l 315 399 m 463 1032 b 350 1135 403 1032 350 1071 b 463 1239 350 1200 403 1239 b 575 1135 521 1239 575 1200 b 463 1032 575 1071 521 1032 m 463 1089 b 510 1135 490 1089 510 1108 b 463 1183 510 1164 490 1183 b 415 1135 435 1183 415 1164 b 463 1089 415 1108 435 1089 "},"Æ":{"ha":1190,"x_min":0,"x_max":0,"o":"m 17 119 l 32 119 b 79 150 57 119 69 125 l 408 981 l 1031 981 b 1119 892 1093 981 1119 954 l 1119 768 l 993 768 l 993 831 b 963 861 993 851 982 861 l 692 861 l 692 554 l 1026 554 l 1026 435 l 692 435 l 692 150 b 728 119 692 126 704 119 l 1000 119 b 1031 150 1019 119 1031 131 l 1031 213 l 1158 213 l 1158 89 b 1068 0 1158 26 1132 0 l 650 0 b 551 89 581 0 551 19 l 551 435 l 331 435 l 199 81 b 83 0 175 17 153 0 l 17 0 m 551 554 l 551 861 l 490 861 l 376 554 "},"Ç":{"ha":974,"x_min":0,"x_max":0,"o":"m 47 499 b 539 997 47 782 258 997 b 907 785 671 997 907 942 l 907 689 l 779 689 l 779 744 b 542 868 779 826 640 868 b 192 499 344 868 192 719 b 542 114 192 285 342 114 b 856 246 738 114 856 246 l 931 147 b 556 -17 931 147 799 -10 l 544 -69 b 644 -171 601 -76 644 -111 b 503 -287 644 -257 574 -287 b 425 -278 458 -287 425 -278 l 425 -197 b 485 -207 425 -197 453 -207 b 550 -164 519 -207 550 -196 b 478 -121 550 -135 521 -121 b 453 -122 465 -121 453 -122 l 481 -14 b 47 499 219 14 47 229 "},"È":{"ha":789,"x_min":0,"x_max":0,"o":"m 143 861 l 53 861 l 53 981 l 629 981 b 718 892 692 981 718 954 l 718 768 l 592 768 l 592 831 b 561 861 592 851 581 861 l 282 861 l 282 554 l 625 554 l 625 435 l 282 435 l 282 150 b 313 119 282 131 293 119 l 599 119 b 629 150 618 119 629 131 l 629 213 l 757 213 l 757 89 b 668 0 757 26 731 0 l 232 0 b 143 89 169 0 143 26 m 224 1225 l 360 1225 l 461 1049 l 356 1049 "},"É":{"ha":789,"x_min":0,"x_max":0,"o":"m 143 861 l 53 861 l 53 981 l 629 981 b 718 892 692 981 718 954 l 718 768 l 592 768 l 592 831 b 561 861 592 851 581 861 l 282 861 l 282 554 l 625 554 l 625 435 l 282 435 l 282 150 b 313 119 282 131 293 119 l 599 119 b 629 150 618 119 629 131 l 629 213 l 757 213 l 757 89 b 668 0 757 26 731 0 l 232 0 b 143 89 169 0 143 26 m 457 1225 l 593 1225 l 461 1049 l 356 1049 "},"Ê":{"ha":789,"x_min":0,"x_max":0,"o":"m 143 861 l 53 861 l 53 981 l 629 981 b 718 892 692 981 718 954 l 718 768 l 592 768 l 592 831 b 561 861 592 851 581 861 l 282 861 l 282 554 l 625 554 l 625 435 l 282 435 l 282 150 b 313 119 282 131 293 119 l 599 119 b 629 150 618 119 629 131 l 629 213 l 757 213 l 757 89 b 668 0 757 26 731 0 l 232 0 b 143 89 169 0 143 26 m 349 1225 l 472 1225 l 594 1049 l 485 1049 l 413 1160 l 410 1160 l 336 1049 l 228 1049 "},"Ë":{"ha":789,"x_min":0,"x_max":0,"o":"m 143 861 l 53 861 l 53 981 l 629 981 b 718 892 692 981 718 954 l 718 768 l 592 768 l 592 831 b 561 861 592 851 581 861 l 282 861 l 282 554 l 625 554 l 625 435 l 282 435 l 282 150 b 313 119 282 131 293 119 l 599 119 b 629 150 618 119 629 131 l 629 213 l 757 213 l 757 89 b 668 0 757 26 731 0 l 232 0 b 143 89 169 0 143 26 m 475 1225 l 576 1225 l 576 1085 l 475 1085 m 246 1225 l 349 1225 l 349 1085 l 246 1085 "},"Ì":{"ha":440,"x_min":0,"x_max":0,"o":"m 58 119 l 154 119 l 154 861 l 58 861 l 58 981 l 382 981 l 382 861 l 286 861 l 286 119 l 382 119 l 382 0 l 58 0 m 33 1225 l 169 1225 l 271 1049 l 165 1049 "},"Í":{"ha":440,"x_min":0,"x_max":0,"o":"m 58 119 l 154 119 l 154 861 l 58 861 l 58 981 l 382 981 l 382 861 l 286 861 l 286 119 l 382 119 l 382 0 l 58 0 m 267 1225 l 403 1225 l 271 1049 l 165 1049 "},"Î":{"ha":440,"x_min":0,"x_max":0,"o":"m 58 119 l 154 119 l 154 861 l 58 861 l 58 981 l 382 981 l 382 861 l 286 861 l 286 119 l 382 119 l 382 0 l 58 0 m 157 1225 l 281 1225 l 403 1049 l 293 1049 l 221 1160 l 218 1160 l 144 1049 l 36 1049 "},"Ï":{"ha":440,"x_min":0,"x_max":0,"o":"m 58 119 l 154 119 l 154 861 l 58 861 l 58 981 l 382 981 l 382 861 l 286 861 l 286 119 l 382 119 l 382 0 l 58 0 m 283 1225 l 385 1225 l 385 1085 l 283 1085 m 54 1225 l 157 1225 l 157 1085 l 54 1085 "},"Ð":{"ha":993,"x_min":0,"x_max":0,"o":"m 143 433 l 75 433 l 75 547 l 143 547 l 143 861 l 53 861 l 53 981 l 450 981 b 946 492 749 981 946 801 b 450 0 946 181 749 0 l 232 0 b 143 89 169 0 143 26 m 282 150 b 313 119 282 131 293 119 l 440 119 b 801 492 658 119 801 250 b 440 861 801 732 657 861 l 282 861 l 282 547 l 482 547 l 482 433 l 282 433 l 282 150 "},"Ñ":{"ha":1063,"x_min":0,"x_max":0,"o":"m 53 119 l 113 119 b 143 150 132 119 143 131 l 143 981 l 269 981 l 706 360 b 788 226 742 307 788 226 l 790 226 b 781 360 790 226 781 304 l 781 892 b 871 981 781 954 807 981 l 1010 981 l 1010 861 l 950 861 b 919 831 931 861 919 851 l 919 0 l 794 0 l 357 621 b 276 754 321 674 276 754 l 274 754 b 282 621 274 754 282 675 l 282 89 b 193 0 282 26 256 0 l 53 0 m 310 1050 b 453 1226 310 1167 358 1226 b 631 1132 549 1226 569 1132 b 686 1222 672 1132 686 1182 l 778 1222 b 635 1046 778 1107 729 1046 b 457 1140 540 1046 518 1140 b 401 1050 415 1140 401 1092 l 310 1050 "},"Ò":{"ha":1101,"x_min":0,"x_max":0,"o":"m 47 497 b 550 997 47 779 267 997 b 1054 497 833 997 1054 779 b 550 -17 1054 208 833 -17 b 47 497 267 -17 47 208 m 192 497 b 550 114 192 281 351 114 b 910 497 749 114 910 281 b 550 868 910 707 749 868 b 192 497 351 868 192 707 m 360 1225 l 496 1225 l 597 1049 l 492 1049 "},"Ó":{"ha":1101,"x_min":0,"x_max":0,"o":"m 47 497 b 550 997 47 779 267 997 b 1054 497 833 997 1054 779 b 550 -17 1054 208 833 -17 b 47 497 267 -17 47 208 m 192 497 b 550 114 192 281 351 114 b 910 497 749 114 910 281 b 550 868 910 707 749 868 b 192 497 351 868 192 707 m 594 1225 l 731 1225 l 599 1049 l 493 1049 "},"Ô":{"ha":1101,"x_min":0,"x_max":0,"o":"m 47 497 b 550 997 47 779 267 997 b 1054 497 833 997 1054 779 b 550 -17 1054 208 833 -17 b 47 497 267 -17 47 208 m 192 497 b 550 114 192 281 351 114 b 910 497 749 114 910 281 b 550 868 910 707 749 868 b 192 497 351 868 192 707 m 485 1225 l 608 1225 l 731 1049 l 621 1049 l 549 1160 l 546 1160 l 472 1049 l 364 1049 "},"Õ":{"ha":1101,"x_min":0,"x_max":0,"o":"m 47 497 b 550 997 47 779 267 997 b 1054 497 833 997 1054 779 b 550 -17 1054 208 833 -17 b 47 497 267 -17 47 208 m 192 497 b 550 114 192 281 351 114 b 910 497 749 114 910 281 b 550 868 910 707 749 868 b 192 497 351 868 192 707 m 311 1050 b 454 1226 311 1167 360 1226 b 632 1132 550 1226 571 1132 b 688 1222 674 1132 688 1182 l 779 1222 b 636 1046 779 1107 731 1046 b 458 1140 542 1046 519 1140 b 403 1050 417 1140 403 1092 l 311 1050 "},"Ö":{"ha":1101,"x_min":0,"x_max":0,"o":"m 47 497 b 550 997 47 779 267 997 b 1054 497 833 997 1054 779 b 550 -17 1054 208 833 -17 b 47 497 267 -17 47 208 m 192 497 b 550 114 192 281 351 114 b 910 497 749 114 910 281 b 550 868 910 707 749 868 b 192 497 351 868 192 707 m 611 1225 l 713 1225 l 713 1085 l 611 1085 m 382 1225 l 485 1225 l 485 1085 l 382 1085 "},"×":{"ha":851,"x_min":0,"x_max":0,"o":"m 351 396 l 65 715 l 142 792 l 425 475 l 710 792 l 786 715 l 499 396 l 786 76 l 710 0 l 425 315 l 142 0 l 65 76 "},"Ø":{"ha":1104,"x_min":0,"x_max":0,"o":"m 222 104 b 49 497 115 199 49 338 b 551 997 49 779 269 997 b 813 926 649 997 738 971 l 878 1017 l 946 969 l 882 879 b 1056 497 988 788 1056 651 b 551 -17 1056 208 835 -17 b 290 57 456 -17 365 10 l 225 -35 l 156 13 m 735 817 b 551 868 681 849 618 868 b 193 497 353 868 193 707 b 304 218 193 385 236 288 m 368 165 b 551 114 422 132 485 114 b 911 497 750 114 911 281 b 799 764 911 604 867 699 l 368 165 "},"Ù":{"ha":1011,"x_min":0,"x_max":0,"o":"m 132 831 b 101 861 132 851 121 861 l 42 861 l 42 981 l 182 981 b 271 892 244 981 271 954 l 271 349 b 506 114 271 201 364 114 b 740 350 647 114 740 201 l 740 892 b 831 981 740 954 768 981 l 969 981 l 969 861 l 910 861 b 879 831 890 861 879 851 l 879 343 b 507 -17 879 129 731 -17 b 132 343 282 -17 132 129 m 321 1225 l 457 1225 l 558 1049 l 453 1049 "},"Ú":{"ha":1011,"x_min":0,"x_max":0,"o":"m 132 831 b 101 861 132 851 121 861 l 42 861 l 42 981 l 182 981 b 271 892 244 981 271 954 l 271 349 b 506 114 271 201 364 114 b 740 350 647 114 740 201 l 740 892 b 831 981 740 954 768 981 l 969 981 l 969 861 l 910 861 b 879 831 890 861 879 851 l 879 343 b 507 -17 879 129 731 -17 b 132 343 282 -17 132 129 m 554 1225 l 690 1225 l 558 1049 l 453 1049 "},"Û":{"ha":1011,"x_min":0,"x_max":0,"o":"m 132 831 b 101 861 132 851 121 861 l 42 861 l 42 981 l 182 981 b 271 892 244 981 271 954 l 271 349 b 506 114 271 201 364 114 b 740 350 647 114 740 201 l 740 892 b 831 981 740 954 768 981 l 969 981 l 969 861 l 910 861 b 879 831 890 861 879 851 l 879 343 b 507 -17 879 129 731 -17 b 132 343 282 -17 132 129 m 444 1225 l 568 1225 l 690 1049 l 581 1049 l 508 1160 l 506 1160 l 432 1049 l 324 1049 "},"Ü":{"ha":1011,"x_min":0,"x_max":0,"o":"m 132 831 b 101 861 132 851 121 861 l 42 861 l 42 981 l 182 981 b 271 892 244 981 271 954 l 271 349 b 506 114 271 201 364 114 b 740 350 647 114 740 201 l 740 892 b 831 981 740 954 768 981 l 969 981 l 969 861 l 910 861 b 879 831 890 861 879 851 l 879 343 b 507 -17 879 129 731 -17 b 132 343 282 -17 132 129 m 571 1225 l 672 1225 l 672 1085 l 571 1085 m 342 1225 l 444 1225 l 444 1085 l 342 1085 "},"Ý":{"ha":838,"x_min":0,"x_max":0,"o":"m 350 428 l 101 826 b 42 861 83 856 71 861 l 14 861 l 14 981 l 78 981 b 210 906 144 981 171 971 l 358 658 b 418 549 390 604 418 549 l 421 549 b 481 658 421 549 447 604 l 628 906 b 760 981 667 971 694 981 l 824 981 l 824 861 l 797 861 b 736 826 768 861 754 856 l 489 428 l 489 0 l 350 0 m 467 1225 l 603 1225 l 471 1049 l 365 1049 "},"Þ":{"ha":828,"x_min":0,"x_max":0,"o":"m 143 831 b 113 861 143 851 132 861 l 53 861 l 53 981 l 193 981 b 282 892 256 981 282 954 l 282 814 l 485 814 b 792 507 664 814 792 693 b 483 196 792 321 664 196 l 282 196 l 282 0 l 143 0 m 463 315 b 649 507 578 315 649 389 b 464 694 649 625 578 694 l 282 694 l 282 315 "},"ß":{"ha":803,"x_min":0,"x_max":0,"o":"m 26 115 l 79 115 b 110 146 99 115 110 125 l 110 736 b 410 997 110 907 258 997 b 674 774 564 997 674 897 b 535 522 674 643 535 604 b 783 194 535 429 783 379 b 546 -17 783 53 668 -17 b 335 53 411 -17 335 53 l 372 157 b 529 100 372 157 443 100 b 646 200 590 100 646 128 b 397 518 646 318 397 364 b 538 772 397 631 538 674 b 406 881 538 826 493 881 b 244 726 324 881 244 831 l 244 89 b 154 0 244 26 218 0 l 26 0 "},"à":{"ha":756,"x_min":0,"x_max":0,"o":"m 51 197 b 475 438 51 424 357 438 l 506 438 l 506 444 b 350 613 506 568 460 613 b 235 558 317 613 235 604 l 235 518 l 108 518 l 108 588 b 351 719 108 703 283 719 b 640 444 579 719 640 600 l 640 146 b 671 115 640 125 651 115 l 728 115 l 728 0 l 601 0 b 514 81 539 0 514 31 b 517 122 514 104 517 122 l 514 122 b 290 -17 514 122 458 -17 b 51 197 169 -17 51 57 m 188 204 b 319 92 188 147 231 92 b 507 315 433 92 507 211 l 507 339 l 472 339 b 188 204 371 339 188 333 m 182 981 l 318 981 l 419 804 l 314 804 "},"á":{"ha":756,"x_min":0,"x_max":0,"o":"m 51 197 b 475 438 51 424 357 438 l 506 438 l 506 444 b 350 613 506 568 460 613 b 235 558 317 613 235 604 l 235 518 l 108 518 l 108 588 b 351 719 108 703 283 719 b 640 444 579 719 640 600 l 640 146 b 671 115 640 125 651 115 l 728 115 l 728 0 l 601 0 b 514 81 539 0 514 31 b 517 122 514 104 517 122 l 514 122 b 290 -17 514 122 458 -17 b 51 197 169 -17 51 57 m 188 204 b 319 92 188 147 231 92 b 507 315 433 92 507 211 l 507 339 l 472 339 b 188 204 371 339 188 333 m 415 981 l 551 981 l 419 804 l 314 804 "},"â":{"ha":756,"x_min":0,"x_max":0,"o":"m 51 197 b 475 438 51 424 357 438 l 506 438 l 506 444 b 350 613 506 568 460 613 b 235 558 317 613 235 604 l 235 518 l 108 518 l 108 588 b 351 719 108 703 283 719 b 640 444 579 719 640 600 l 640 146 b 671 115 640 125 651 115 l 728 115 l 728 0 l 601 0 b 514 81 539 0 514 31 b 517 122 514 104 517 122 l 514 122 b 290 -17 514 122 458 -17 b 51 197 169 -17 51 57 m 188 204 b 319 92 188 147 231 92 b 507 315 433 92 507 211 l 507 339 l 472 339 b 188 204 371 339 188 333 m 307 981 l 431 981 l 553 804 l 443 804 l 371 915 l 368 915 l 294 804 l 186 804 "},"ã":{"ha":756,"x_min":0,"x_max":0,"o":"m 51 197 b 475 438 51 424 357 438 l 506 438 l 506 444 b 350 613 506 568 460 613 b 235 558 317 613 235 604 l 235 518 l 108 518 l 108 588 b 351 719 108 703 283 719 b 640 444 579 719 640 600 l 640 146 b 671 115 640 125 651 115 l 728 115 l 728 0 l 601 0 b 514 81 539 0 514 31 b 517 122 514 104 517 122 l 514 122 b 290 -17 514 122 458 -17 b 51 197 169 -17 51 57 m 188 204 b 319 92 188 147 231 92 b 507 315 433 92 507 211 l 507 339 l 472 339 b 188 204 371 339 188 333 m 133 806 b 276 982 133 922 182 982 b 454 888 372 982 393 888 b 510 978 496 888 510 938 l 601 978 b 458 801 601 863 553 801 b 281 896 364 801 342 896 b 225 806 239 896 225 847 l 133 806 "},"ä":{"ha":756,"x_min":0,"x_max":0,"o":"m 51 197 b 475 438 51 424 357 438 l 506 438 l 506 444 b 350 613 506 568 460 613 b 235 558 317 613 235 604 l 235 518 l 108 518 l 108 588 b 351 719 108 703 283 719 b 640 444 579 719 640 600 l 640 146 b 671 115 640 125 651 115 l 728 115 l 728 0 l 601 0 b 514 81 539 0 514 31 b 517 122 514 104 517 122 l 514 122 b 290 -17 514 122 458 -17 b 51 197 169 -17 51 57 m 188 204 b 319 92 188 147 231 92 b 507 315 433 92 507 211 l 507 339 l 472 339 b 188 204 371 339 188 333 m 433 981 l 535 981 l 535 840 l 433 840 m 204 981 l 307 981 l 307 840 l 204 840 "},"å":{"ha":756,"x_min":0,"x_max":0,"o":"m 51 197 b 475 438 51 424 357 438 l 506 438 l 506 444 b 350 613 506 568 460 613 b 235 558 317 613 235 604 l 235 518 l 108 518 l 108 588 b 351 719 108 703 283 719 b 640 444 579 719 640 600 l 640 146 b 671 115 640 125 651 115 l 728 115 l 728 0 l 601 0 b 514 81 539 0 514 31 b 517 122 514 104 517 122 l 514 122 b 290 -17 514 122 458 -17 b 51 197 169 -17 51 57 m 188 204 b 319 92 188 147 231 92 b 507 315 433 92 507 211 l 507 339 l 472 339 b 188 204 371 339 188 333 m 369 788 b 257 890 310 788 257 826 b 369 994 257 956 310 994 b 482 890 428 994 482 956 b 369 788 482 826 428 788 m 369 844 b 417 890 397 844 417 864 b 369 939 417 919 397 939 b 322 890 342 939 322 919 b 369 844 322 864 342 844 "},"æ":{"ha":1206,"x_min":0,"x_max":0,"o":"m 51 197 b 471 433 51 421 326 433 l 506 433 l 506 446 b 350 613 506 571 456 613 b 235 558 315 613 235 604 l 235 518 l 108 518 l 108 588 b 351 719 108 689 247 719 b 590 596 490 719 557 672 l 593 596 b 843 719 649 678 736 719 b 1142 397 1035 719 1142 574 b 1139 335 1142 369 1139 335 l 636 335 b 869 100 644 181 743 100 b 1075 182 990 100 1075 182 l 1133 86 b 863 -17 1133 86 1029 -17 b 560 160 721 -17 615 50 l 557 160 b 292 -17 544 124 474 -17 b 51 197 163 -17 51 61 m 1006 433 b 840 613 1001 550 926 613 b 642 433 738 613 661 549 m 189 204 b 321 92 189 146 232 92 b 507 317 436 92 507 213 l 507 335 l 442 335 b 189 204 336 335 189 321 "},"ç":{"ha":740,"x_min":0,"x_max":0,"o":"m 50 351 b 428 719 50 558 201 719 b 681 565 513 719 681 685 l 681 493 l 558 493 l 558 532 b 428 604 558 582 479 604 b 188 353 290 604 188 501 b 435 103 188 193 304 103 b 653 200 567 103 653 200 l 711 104 b 426 -17 711 104 613 -15 l 415 -69 b 515 -171 472 -76 515 -111 b 374 -287 515 -257 444 -287 b 296 -278 329 -287 296 -278 l 296 -197 b 356 -207 296 -197 324 -207 b 421 -164 390 -207 421 -196 b 349 -121 421 -135 392 -121 b 324 -122 336 -121 324 -122 l 351 -10 b 50 351 172 19 50 164 "},"è":{"ha":767,"x_min":0,"x_max":0,"o":"m 50 351 b 403 719 50 575 203 719 b 706 390 600 719 706 572 b 701 335 706 372 701 335 l 189 335 b 431 100 197 185 306 100 b 636 182 551 100 636 182 l 694 86 b 424 -17 694 86 590 -17 b 50 351 201 -17 50 143 m 567 433 b 401 613 563 550 489 613 b 194 433 299 613 215 550 m 211 981 l 347 981 l 449 804 l 343 804 "},"é":{"ha":767,"x_min":0,"x_max":0,"o":"m 50 351 b 403 719 50 575 203 719 b 706 390 600 719 706 572 b 701 335 706 372 701 335 l 189 335 b 431 100 197 185 306 100 b 636 182 551 100 636 182 l 694 86 b 424 -17 694 86 590 -17 b 50 351 201 -17 50 143 m 567 433 b 401 613 563 550 489 613 b 194 433 299 613 215 550 m 446 981 l 582 981 l 450 804 l 344 804 "},"ê":{"ha":767,"x_min":0,"x_max":0,"o":"m 50 351 b 403 719 50 575 203 719 b 706 390 600 719 706 572 b 701 335 706 372 701 335 l 189 335 b 431 100 197 185 306 100 b 636 182 551 100 636 182 l 694 86 b 424 -17 694 86 590 -17 b 50 351 201 -17 50 143 m 567 433 b 401 613 563 550 489 613 b 194 433 299 613 215 550 m 336 981 l 460 981 l 582 804 l 472 804 l 400 915 l 397 915 l 324 804 l 215 804 "},"ë":{"ha":767,"x_min":0,"x_max":0,"o":"m 50 351 b 403 719 50 575 203 719 b 706 390 600 719 706 572 b 701 335 706 372 701 335 l 189 335 b 431 100 197 185 306 100 b 636 182 551 100 636 182 l 694 86 b 424 -17 694 86 590 -17 b 50 351 201 -17 50 143 m 567 433 b 401 613 563 550 489 613 b 194 433 299 613 215 550 m 463 981 l 564 981 l 564 840 l 463 840 m 233 981 l 336 981 l 336 840 l 233 840 "},"ì":{"ha":385,"x_min":0,"x_max":0,"o":"m 125 557 b 94 588 125 578 114 588 l 38 588 l 38 703 l 171 703 b 260 614 233 703 260 676 l 260 146 b 290 115 260 125 271 115 l 347 115 l 347 0 l 215 0 b 125 89 151 0 125 26 m -4 981 l 132 981 l 233 804 l 128 804 "},"í":{"ha":385,"x_min":0,"x_max":0,"o":"m 125 557 b 94 588 125 578 114 588 l 38 588 l 38 703 l 171 703 b 260 614 233 703 260 676 l 260 146 b 290 115 260 125 271 115 l 347 115 l 347 0 l 215 0 b 125 89 151 0 125 26 m 229 981 l 365 981 l 233 804 l 128 804 "},"î":{"ha":385,"x_min":0,"x_max":0,"o":"m 125 557 b 94 588 125 578 114 588 l 38 588 l 38 703 l 171 703 b 260 614 233 703 260 676 l 260 146 b 290 115 260 125 271 115 l 347 115 l 347 0 l 215 0 b 125 89 151 0 125 26 m 121 981 l 244 981 l 367 804 l 256 804 l 183 917 l 181 917 l 108 804 l -1 804 "},"ï":{"ha":385,"x_min":0,"x_max":0,"o":"m 125 557 b 94 588 125 578 114 588 l 38 588 l 38 703 l 171 703 b 260 614 233 703 260 676 l 260 146 b 290 115 260 125 271 115 l 347 115 l 347 0 l 215 0 b 125 89 151 0 125 26 m 225 981 l 328 981 l 328 840 l 225 840 m 32 981 l 133 981 l 133 840 l 32 840 "},"ð":{"ha":799,"x_min":0,"x_max":0,"o":"m 53 326 b 382 647 53 490 168 647 b 553 589 493 647 553 589 l 556 589 b 404 774 526 667 472 728 l 153 663 l 139 749 l 311 825 b 132 889 256 853 193 875 l 169 999 b 457 888 268 976 368 940 l 650 972 l 663 888 l 535 829 b 735 382 651 732 735 585 b 382 -17 735 194 635 -17 b 53 326 171 -17 53 153 m 189 324 b 385 100 189 210 256 100 b 597 358 536 100 597 239 b 401 539 597 463 521 539 b 189 324 257 539 189 432 "},"ñ":{"ha":869,"x_min":0,"x_max":0,"o":"m 118 557 b 88 588 118 578 107 588 l 31 588 l 31 703 l 160 703 b 249 625 219 703 249 676 l 249 599 b 246 558 249 576 246 558 l 249 558 b 513 719 276 619 360 719 b 754 450 679 719 754 629 l 754 146 b 785 115 754 125 765 115 l 842 115 l 842 0 l 708 0 b 619 89 646 0 619 26 l 619 419 b 488 596 619 518 600 596 b 265 424 379 596 293 525 b 253 329 256 396 253 364 l 253 0 l 118 0 m 213 806 b 356 982 213 922 261 982 b 533 888 451 982 472 888 b 589 978 575 888 589 938 l 681 978 b 538 801 681 863 632 801 b 360 896 443 801 421 896 b 304 806 318 896 304 847 l 213 806 "},"ò":{"ha":857,"x_min":0,"x_max":0,"o":"m 50 353 b 428 719 50 565 218 719 b 807 353 638 719 807 565 b 428 -17 807 139 638 -17 b 50 353 218 -17 50 139 m 188 353 b 428 100 188 206 296 100 b 669 353 561 100 669 206 b 428 603 669 499 561 603 b 188 353 296 603 188 499 m 242 981 l 378 981 l 479 804 l 374 804 "},"ó":{"ha":857,"x_min":0,"x_max":0,"o":"m 50 353 b 428 719 50 565 218 719 b 807 353 638 719 807 565 b 428 -17 807 139 638 -17 b 50 353 218 -17 50 139 m 188 353 b 428 100 188 206 296 100 b 669 353 561 100 669 206 b 428 603 669 499 561 603 b 188 353 296 603 188 499 m 475 981 l 611 981 l 479 804 l 374 804 "},"ô":{"ha":857,"x_min":0,"x_max":0,"o":"m 50 353 b 428 719 50 565 218 719 b 807 353 638 719 807 565 b 428 -17 807 139 638 -17 b 50 353 218 -17 50 139 m 188 353 b 428 100 188 206 296 100 b 669 353 561 100 669 206 b 428 603 669 499 561 603 b 188 353 296 603 188 499 m 367 981 l 490 981 l 613 804 l 503 804 l 431 915 l 428 915 l 354 804 l 246 804 "},"õ":{"ha":857,"x_min":0,"x_max":0,"o":"m 50 353 b 428 719 50 565 218 719 b 807 353 638 719 807 565 b 428 -17 807 139 638 -17 b 50 353 218 -17 50 139 m 188 353 b 428 100 188 206 296 100 b 669 353 561 100 669 206 b 428 603 669 499 561 603 b 188 353 296 603 188 499 m 193 806 b 336 982 193 922 242 982 b 514 888 432 982 453 888 b 569 978 556 888 569 938 l 661 978 b 518 801 661 863 613 801 b 340 896 424 801 401 896 b 285 806 299 896 285 847 l 193 806 "},"ö":{"ha":857,"x_min":0,"x_max":0,"o":"m 50 353 b 428 719 50 565 218 719 b 807 353 638 719 807 565 b 428 -17 807 139 638 -17 b 50 353 218 -17 50 139 m 188 353 b 428 100 188 206 296 100 b 669 353 561 100 669 206 b 428 603 669 499 561 603 b 188 353 296 603 188 499 m 493 981 l 594 981 l 594 840 l 493 840 m 264 981 l 367 981 l 367 840 l 264 840 "},"÷":{"ha":851,"x_min":0,"x_max":0,"o":"m 351 731 l 482 731 l 482 604 l 351 604 m 67 450 l 768 450 l 768 342 l 67 342 m 351 188 l 482 188 l 482 61 l 351 61 "},"ø":{"ha":857,"x_min":0,"x_max":0,"o":"m 183 68 b 50 353 101 133 50 235 b 428 719 50 565 218 719 b 617 672 496 719 561 703 l 672 750 l 738 703 l 683 628 b 807 353 758 561 807 465 b 428 -17 807 139 638 -17 b 251 25 364 -17 304 -3 l 197 -53 l 131 -4 m 544 572 b 428 603 510 592 471 603 b 188 353 296 603 188 499 b 258 172 188 279 214 218 m 324 125 b 428 100 356 108 390 100 b 669 353 561 100 669 206 b 606 519 669 419 646 478 l 324 125 "},"ù":{"ha":863,"x_min":0,"x_max":0,"o":"m 115 557 b 85 588 115 578 104 588 l 28 588 l 28 703 l 160 703 b 250 617 224 703 250 676 l 250 283 b 381 108 250 186 269 108 b 610 374 525 108 610 235 l 610 703 l 744 703 l 744 146 b 775 115 744 125 756 115 l 832 115 l 832 0 l 703 0 b 614 78 643 0 614 28 l 614 104 b 615 144 614 126 615 144 l 613 144 b 354 -17 583 75 493 -17 b 115 253 196 -17 115 67 m 236 981 l 372 981 l 474 804 l 368 804 "},"ú":{"ha":863,"x_min":0,"x_max":0,"o":"m 115 557 b 85 588 115 578 104 588 l 28 588 l 28 703 l 160 703 b 250 617 224 703 250 676 l 250 283 b 381 108 250 186 269 108 b 610 374 525 108 610 235 l 610 703 l 744 703 l 744 146 b 775 115 744 125 756 115 l 832 115 l 832 0 l 703 0 b 614 78 643 0 614 28 l 614 104 b 615 144 614 126 615 144 l 613 144 b 354 -17 583 75 493 -17 b 115 253 196 -17 115 67 m 469 981 l 606 981 l 474 804 l 368 804 "},"û":{"ha":863,"x_min":0,"x_max":0,"o":"m 115 557 b 85 588 115 578 104 588 l 28 588 l 28 703 l 160 703 b 250 617 224 703 250 676 l 250 283 b 381 108 250 186 269 108 b 610 374 525 108 610 235 l 610 703 l 744 703 l 744 146 b 775 115 744 125 756 115 l 832 115 l 832 0 l 703 0 b 614 78 643 0 614 28 l 614 104 b 615 144 614 126 615 144 l 613 144 b 354 -17 583 75 493 -17 b 115 253 196 -17 115 67 m 360 981 l 483 981 l 606 804 l 496 804 l 424 915 l 421 915 l 347 804 l 239 804 "},"ü":{"ha":863,"x_min":0,"x_max":0,"o":"m 115 557 b 85 588 115 578 104 588 l 28 588 l 28 703 l 160 703 b 250 617 224 703 250 676 l 250 283 b 381 108 250 186 269 108 b 610 374 525 108 610 235 l 610 703 l 744 703 l 744 146 b 775 115 744 125 756 115 l 832 115 l 832 0 l 703 0 b 614 78 643 0 614 28 l 614 104 b 615 144 614 126 615 144 l 613 144 b 354 -17 583 75 493 -17 b 115 253 196 -17 115 67 m 486 981 l 588 981 l 588 840 l 486 840 m 257 981 l 360 981 l 360 840 l 257 840 "},"ý":{"ha":758,"x_min":0,"x_max":0,"o":"m 63 -125 b 165 -175 63 -126 106 -175 b 289 -74 219 -175 263 -136 l 321 3 l 81 560 b 36 588 71 582 58 588 l 19 588 l 19 703 l 99 703 b 199 632 156 703 176 688 l 358 236 b 383 151 372 200 383 151 l 386 151 b 410 236 386 151 396 200 l 556 632 b 658 703 576 688 601 703 l 739 703 l 739 588 l 721 588 b 676 560 699 588 686 582 l 399 -126 b 172 -292 356 -235 271 -292 b 7 -222 68 -292 7 -222 l 63 -125 m 426 981 l 563 981 l 431 804 l 325 804 "},"þ":{"ha":835,"x_min":0,"x_max":0,"o":"m 114 835 b 83 865 114 856 103 865 l 26 865 l 26 981 l 160 981 b 249 892 222 981 249 954 l 249 665 b 246 604 249 629 246 604 l 249 604 b 469 719 290 676 360 719 b 783 351 663 719 783 567 b 469 -17 783 138 663 -17 b 249 96 304 -17 249 96 l 246 96 b 249 35 246 96 249 71 l 249 -278 l 114 -278 m 244 351 b 449 100 244 167 343 100 b 646 351 568 100 646 203 b 449 603 646 500 568 603 b 244 351 339 603 244 536 "},"ÿ":{"ha":758,"x_min":0,"x_max":0,"o":"m 63 -125 b 165 -175 63 -126 106 -175 b 289 -74 219 -175 263 -136 l 321 3 l 81 560 b 36 588 71 582 58 588 l 19 588 l 19 703 l 99 703 b 199 632 156 703 176 688 l 358 236 b 383 151 372 200 383 151 l 386 151 b 410 236 386 151 396 200 l 556 632 b 658 703 576 688 601 703 l 739 703 l 739 588 l 721 588 b 676 560 699 588 686 582 l 399 -126 b 172 -292 356 -235 271 -292 b 7 -222 68 -292 7 -222 l 63 -125 m 443 981 l 544 981 l 544 840 l 443 840 m 214 981 l 317 981 l 317 840 l 214 840 "},"Ā":{"ha":924,"x_min":0,"x_max":0,"o":"m 17 119 l 36 119 b 82 150 61 119 74 125 l 389 981 l 535 981 l 842 150 b 888 119 850 125 864 119 l 908 119 l 908 0 l 831 0 b 717 81 761 0 739 17 l 643 282 l 282 282 l 208 81 b 93 0 185 17 163 0 l 17 0 m 608 399 l 501 697 b 463 835 483 749 463 835 l 460 835 b 422 697 460 835 440 749 l 315 399 m 286 1183 l 639 1183 l 639 1089 l 286 1089 "},"ā":{"ha":756,"x_min":0,"x_max":0,"o":"m 51 197 b 475 438 51 424 357 438 l 506 438 l 506 444 b 350 613 506 568 460 613 b 235 558 317 613 235 604 l 235 518 l 108 518 l 108 588 b 351 719 108 703 283 719 b 640 444 579 719 640 600 l 640 146 b 671 115 640 125 651 115 l 728 115 l 728 0 l 601 0 b 514 81 539 0 514 31 b 517 122 514 104 517 122 l 514 122 b 290 -17 514 122 458 -17 b 51 197 169 -17 51 57 m 188 204 b 319 92 188 147 231 92 b 507 315 433 92 507 211 l 507 339 l 472 339 b 188 204 371 339 188 333 m 193 939 l 546 939 l 546 844 l 193 844 "},"Ă":{"ha":924,"x_min":0,"x_max":0,"o":"m 17 119 l 36 119 b 82 150 61 119 74 125 l 389 981 l 535 981 l 842 150 b 888 119 850 125 864 119 l 908 119 l 908 0 l 831 0 b 717 81 761 0 739 17 l 643 282 l 282 282 l 208 81 b 93 0 185 17 163 0 l 17 0 m 608 399 l 501 697 b 463 835 483 749 463 835 l 460 835 b 422 697 460 835 440 749 l 315 399 m 463 1043 b 285 1206 375 1043 285 1099 l 285 1225 l 378 1225 l 378 1207 b 463 1124 378 1149 419 1124 b 549 1208 506 1124 549 1149 l 549 1225 l 640 1225 l 640 1206 b 463 1043 640 1099 551 1043 "},"ă":{"ha":756,"x_min":0,"x_max":0,"o":"m 51 197 b 475 438 51 424 357 438 l 506 438 l 506 444 b 350 613 506 568 460 613 b 235 558 317 613 235 604 l 235 518 l 108 518 l 108 588 b 351 719 108 703 283 719 b 640 444 579 719 640 600 l 640 146 b 671 115 640 125 651 115 l 728 115 l 728 0 l 601 0 b 514 81 539 0 514 31 b 517 122 514 104 517 122 l 514 122 b 290 -17 514 122 458 -17 b 51 197 169 -17 51 57 m 188 204 b 319 92 188 147 231 92 b 507 315 433 92 507 211 l 507 339 l 472 339 b 188 204 371 339 188 333 m 368 799 b 190 961 281 799 190 854 l 190 981 l 283 981 l 283 963 b 368 879 283 904 325 879 b 454 964 411 879 454 904 l 454 981 l 546 981 l 546 961 b 368 799 546 854 457 799 "},"Ą":{"ha":924,"x_min":0,"x_max":0,"o":"m 17 119 l 36 119 b 82 150 61 119 74 125 l 389 981 l 535 981 l 842 150 b 888 119 850 125 864 119 l 908 119 l 908 0 b 819 -124 879 -24 819 -75 b 896 -183 819 -176 896 -183 l 864 -287 b 703 -156 864 -287 703 -275 b 797 1 703 -82 758 -28 l 797 4 b 717 81 751 10 735 29 l 643 282 l 282 282 l 208 81 b 93 0 185 17 163 0 l 17 0 m 608 399 l 501 697 b 463 835 483 749 463 835 l 460 835 b 422 697 460 835 440 749 l 315 399 "},"ą":{"ha":756,"x_min":0,"x_max":0,"o":"m 51 197 b 476 438 51 424 357 438 l 506 438 l 506 444 b 350 613 506 568 460 613 b 235 558 315 613 235 603 l 235 518 l 108 518 l 108 588 b 351 719 108 701 283 719 b 640 444 579 719 640 600 l 640 146 b 671 115 640 125 651 115 l 728 115 l 728 0 b 639 -124 700 -24 639 -75 b 715 -183 639 -176 715 -183 l 685 -287 b 522 -156 685 -287 522 -275 b 614 -3 522 -86 575 -33 l 614 0 l 590 0 b 514 76 531 0 514 40 l 514 83 b 517 122 514 104 517 122 l 514 122 b 286 -17 514 122 461 -17 b 51 197 160 -17 51 68 m 188 204 b 319 92 188 147 231 92 b 507 315 433 92 507 211 l 507 339 l 472 339 b 188 204 371 339 188 333 "},"Ć":{"ha":974,"x_min":0,"x_max":0,"o":"m 47 499 b 539 997 47 782 258 997 b 907 785 671 997 907 942 l 907 689 l 779 689 l 779 744 b 542 868 779 826 640 868 b 192 499 344 868 192 719 b 546 114 192 285 346 114 b 856 247 740 114 856 247 l 931 147 b 544 -17 931 147 792 -17 b 47 499 251 -17 47 210 m 581 1225 l 717 1225 l 585 1049 l 479 1049 "},"ć":{"ha":740,"x_min":0,"x_max":0,"o":"m 50 351 b 428 719 50 558 201 719 b 681 565 513 719 681 685 l 681 493 l 558 493 l 558 532 b 428 604 558 582 479 604 b 188 353 290 604 188 501 b 435 103 188 193 304 103 b 653 200 567 103 653 200 l 711 104 b 425 -17 711 104 611 -17 b 50 351 204 -17 50 140 m 451 981 l 588 981 l 456 804 l 350 804 "},"Ĉ":{"ha":974,"x_min":0,"x_max":0,"o":"m 47 499 b 539 997 47 782 258 997 b 907 785 671 997 907 942 l 907 689 l 779 689 l 779 744 b 542 868 779 826 640 868 b 192 499 344 868 192 719 b 546 114 192 285 346 114 b 856 247 740 114 856 247 l 931 147 b 544 -17 931 147 792 -17 b 47 499 251 -17 47 210 m 472 1225 l 596 1225 l 718 1049 l 608 1049 l 536 1160 l 533 1160 l 460 1049 l 351 1049 "},"ĉ":{"ha":740,"x_min":0,"x_max":0,"o":"m 50 351 b 428 719 50 558 201 719 b 681 565 513 719 681 685 l 681 493 l 558 493 l 558 532 b 428 604 558 582 479 604 b 188 353 290 604 188 501 b 435 103 188 193 304 103 b 653 200 567 103 653 200 l 711 104 b 425 -17 711 104 611 -17 b 50 351 204 -17 50 140 m 343 981 l 467 981 l 589 804 l 479 804 l 407 915 l 404 915 l 331 804 l 222 804 "},"Ċ":{"ha":974,"x_min":0,"x_max":0,"o":"m 47 499 b 539 997 47 782 258 997 b 907 785 671 997 907 942 l 907 689 l 779 689 l 779 744 b 542 868 779 826 640 868 b 192 499 344 868 192 719 b 546 114 192 285 346 114 b 856 247 740 114 856 247 l 931 147 b 544 -17 931 147 792 -17 b 47 499 251 -17 47 210 m 474 1225 l 594 1225 l 594 1085 l 474 1085 "},"ċ":{"ha":740,"x_min":0,"x_max":0,"o":"m 50 351 b 428 719 50 558 201 719 b 681 565 513 719 681 685 l 681 493 l 558 493 l 558 532 b 428 604 558 582 479 604 b 188 353 290 604 188 501 b 435 103 188 193 304 103 b 653 200 567 103 653 200 l 711 104 b 425 -17 711 104 611 -17 b 50 351 204 -17 50 140 m 344 981 l 465 981 l 465 840 l 344 840 "},"Č":{"ha":974,"x_min":0,"x_max":0,"o":"m 47 499 b 539 997 47 782 258 997 b 907 785 671 997 907 942 l 907 689 l 779 689 l 779 744 b 542 868 779 826 640 868 b 192 499 344 868 192 719 b 546 114 192 285 346 114 b 856 247 740 114 856 247 l 931 147 b 544 -17 931 147 792 -17 b 47 499 251 -17 47 210 m 350 1225 l 460 1225 l 532 1113 l 535 1113 l 607 1225 l 718 1225 l 596 1049 l 472 1049 "},"č":{"ha":740,"x_min":0,"x_max":0,"o":"m 50 351 b 428 719 50 558 201 719 b 681 565 513 719 681 685 l 681 493 l 558 493 l 558 532 b 428 604 558 582 479 604 b 188 353 290 604 188 501 b 435 103 188 193 304 103 b 653 200 567 103 653 200 l 711 104 b 425 -17 711 104 611 -17 b 50 351 204 -17 50 140 m 221 981 l 331 981 l 403 868 l 406 868 l 478 981 l 589 981 l 467 804 l 343 804 "},"Ď":{"ha":993,"x_min":0,"x_max":0,"o":"m 143 861 l 53 861 l 53 981 l 450 981 b 946 492 749 981 946 801 b 450 0 946 181 749 0 l 232 0 b 143 89 169 0 143 26 m 282 150 b 313 119 282 131 293 119 l 440 119 b 801 492 658 119 801 250 b 440 861 801 732 657 861 l 282 861 l 282 150 m 292 1225 l 401 1225 l 474 1113 l 476 1113 l 549 1225 l 660 1225 l 538 1049 l 414 1049 "},"ď":{"ha":960,"x_min":0,"x_max":0,"o":"m 51 351 b 369 719 51 575 183 719 b 586 611 532 719 586 611 l 589 611 b 586 658 589 611 586 632 l 586 835 b 556 865 586 856 575 865 l 499 865 l 499 981 l 632 981 b 721 892 694 981 721 954 l 721 146 b 751 115 721 125 732 115 l 808 115 l 808 0 l 679 0 b 593 75 618 0 593 28 b 594 111 593 96 594 111 l 592 111 b 361 -17 592 111 533 -17 b 51 351 171 -17 51 133 m 388 100 b 590 353 492 100 590 174 b 390 603 590 478 525 603 b 189 351 281 603 189 511 b 388 100 189 199 271 100 m 907 981 b 915 913 913 956 915 933 b 882 785 915 869 904 832 l 786 785 b 813 899 803 828 813 863 b 797 981 813 924 808 950 "},"Đ":{"ha":993,"x_min":0,"x_max":0,"o":"m 143 433 l 75 433 l 75 547 l 143 547 l 143 861 l 53 861 l 53 981 l 450 981 b 946 492 749 981 946 801 b 450 0 946 181 749 0 l 232 0 b 143 89 169 0 143 26 m 282 150 b 313 119 282 131 293 119 l 440 119 b 801 492 658 119 801 250 b 440 861 801 732 657 861 l 282 861 l 282 547 l 482 547 l 482 433 l 282 433 l 282 150 "},"đ":{"ha":842,"x_min":0,"x_max":0,"o":"m 361 -17 b 51 351 171 -17 51 133 b 369 719 51 575 183 719 b 586 611 532 719 586 611 l 589 611 b 586 658 589 611 586 632 l 586 782 l 407 782 l 407 869 l 586 869 l 586 981 l 721 981 l 721 869 l 828 869 l 828 782 l 721 782 l 721 146 b 751 115 721 125 732 115 l 808 115 l 808 0 l 679 0 b 593 75 618 0 593 28 b 594 111 593 96 594 111 l 592 111 b 361 -17 592 111 533 -17 m 388 100 b 590 353 492 100 590 174 b 390 603 590 478 525 603 b 189 351 281 603 189 511 b 388 100 189 199 271 100 "},"Ē":{"ha":789,"x_min":0,"x_max":0,"o":"m 143 861 l 53 861 l 53 981 l 629 981 b 718 892 692 981 718 954 l 718 768 l 592 768 l 592 831 b 561 861 592 851 581 861 l 282 861 l 282 554 l 625 554 l 625 435 l 282 435 l 282 150 b 313 119 282 131 293 119 l 599 119 b 629 150 618 119 629 131 l 629 213 l 757 213 l 757 89 b 668 0 757 26 731 0 l 232 0 b 143 89 169 0 143 26 m 235 1183 l 588 1183 l 588 1089 l 235 1089 "},"ē":{"ha":767,"x_min":0,"x_max":0,"o":"m 50 351 b 403 719 50 575 203 719 b 706 390 600 719 706 572 b 701 335 706 372 701 335 l 189 335 b 431 100 197 185 306 100 b 636 182 551 100 636 182 l 694 86 b 424 -17 694 86 590 -17 b 50 351 201 -17 50 143 m 567 433 b 401 613 563 550 489 613 b 194 433 299 613 215 550 m 222 939 l 575 939 l 575 844 l 222 844 "},"Ĕ":{"ha":789,"x_min":0,"x_max":0,"o":"m 143 861 l 53 861 l 53 981 l 629 981 b 718 892 692 981 718 954 l 718 768 l 592 768 l 592 831 b 561 861 592 851 581 861 l 282 861 l 282 554 l 625 554 l 625 435 l 282 435 l 282 150 b 313 119 282 131 293 119 l 599 119 b 629 150 618 119 629 131 l 629 213 l 757 213 l 757 89 b 668 0 757 26 731 0 l 232 0 b 143 89 169 0 143 26 m 410 1043 b 232 1206 322 1043 232 1099 l 232 1225 l 325 1225 l 325 1207 b 410 1124 325 1149 367 1124 b 496 1208 453 1124 496 1149 l 496 1225 l 588 1225 l 588 1206 b 410 1043 588 1099 499 1043 "},"ĕ":{"ha":767,"x_min":0,"x_max":0,"o":"m 50 351 b 403 719 50 575 203 719 b 706 390 600 719 706 572 b 701 335 706 372 701 335 l 189 335 b 431 100 197 185 306 100 b 636 182 551 100 636 182 l 694 86 b 424 -17 694 86 590 -17 b 50 351 201 -17 50 143 m 567 433 b 401 613 563 550 489 613 b 194 433 299 613 215 550 m 399 799 b 221 961 311 799 221 854 l 221 981 l 314 981 l 314 963 b 399 879 314 904 356 879 b 485 964 442 879 485 904 l 485 981 l 576 981 l 576 961 b 399 799 576 854 488 799 "},"Ė":{"ha":789,"x_min":0,"x_max":0,"o":"m 143 861 l 53 861 l 53 981 l 629 981 b 718 892 692 981 718 954 l 718 768 l 592 768 l 592 831 b 561 861 592 851 581 861 l 282 861 l 282 554 l 625 554 l 625 435 l 282 435 l 282 150 b 313 119 282 131 293 119 l 599 119 b 629 150 618 119 629 131 l 629 213 l 757 213 l 757 89 b 668 0 757 26 731 0 l 232 0 b 143 89 169 0 143 26 m 350 1225 l 471 1225 l 471 1085 l 350 1085 "},"ė":{"ha":767,"x_min":0,"x_max":0,"o":"m 50 351 b 403 719 50 575 203 719 b 706 390 600 719 706 572 b 701 335 706 372 701 335 l 189 335 b 431 100 197 185 306 100 b 636 182 551 100 636 182 l 694 86 b 424 -17 694 86 590 -17 b 50 351 201 -17 50 143 m 567 433 b 401 613 563 550 489 613 b 194 433 299 613 215 550 m 338 981 l 458 981 l 458 840 l 338 840 "},"Ę":{"ha":789,"x_min":0,"x_max":0,"o":"m 143 861 l 53 861 l 53 981 l 629 981 b 718 892 692 981 718 954 l 718 768 l 592 768 l 592 831 b 561 861 592 851 581 861 l 282 861 l 282 554 l 625 554 l 625 435 l 282 435 l 282 150 b 313 119 282 131 293 119 l 599 119 b 629 150 618 119 629 131 l 629 213 l 757 213 l 757 89 b 725 11 757 51 749 31 b 621 -124 683 -25 621 -69 b 697 -183 621 -176 697 -183 l 667 -287 b 504 -156 667 -287 504 -275 b 594 -3 504 -86 554 -35 l 594 0 l 232 0 b 143 89 169 0 143 26 "},"ę":{"ha":767,"x_min":0,"x_max":0,"o":"m 50 351 b 403 719 50 575 203 719 b 706 390 600 719 706 572 b 701 335 706 372 701 335 l 189 335 b 431 100 197 185 306 100 b 636 182 551 100 636 182 l 694 86 b 508 -124 642 35 508 -31 b 586 -183 508 -176 586 -183 l 554 -287 b 393 -156 554 -287 393 -275 b 471 -15 393 -94 433 -49 l 471 -12 b 424 -17 456 -15 440 -17 b 50 351 201 -17 50 143 m 567 433 b 401 613 561 550 490 613 b 194 433 299 613 215 550 "},"Ě":{"ha":789,"x_min":0,"x_max":0,"o":"m 143 861 l 53 861 l 53 981 l 629 981 b 718 892 692 981 718 954 l 718 768 l 592 768 l 592 831 b 561 861 592 851 581 861 l 282 861 l 282 554 l 625 554 l 625 435 l 282 435 l 282 150 b 313 119 282 131 293 119 l 599 119 b 629 150 618 119 629 131 l 629 213 l 757 213 l 757 89 b 668 0 757 26 731 0 l 232 0 b 143 89 169 0 143 26 m 226 1225 l 336 1225 l 408 1113 l 411 1113 l 483 1225 l 594 1225 l 472 1049 l 349 1049 "},"ě":{"ha":767,"x_min":0,"x_max":0,"o":"m 50 351 b 403 719 50 575 203 719 b 706 390 600 719 706 572 b 701 335 706 372 701 335 l 189 335 b 431 100 197 185 306 100 b 636 182 551 100 636 182 l 694 86 b 424 -17 694 86 590 -17 b 50 351 201 -17 50 143 m 567 433 b 401 613 563 550 489 613 b 194 433 299 613 215 550 m 214 981 l 324 981 l 396 868 l 399 868 l 471 981 l 582 981 l 460 804 l 336 804 "},"Ĝ":{"ha":1022,"x_min":0,"x_max":0,"o":"m 49 494 b 543 997 49 776 263 997 b 893 876 778 997 893 876 l 814 775 b 546 869 814 775 708 869 b 193 494 350 869 193 713 b 549 114 193 267 350 114 b 828 236 724 114 828 236 l 828 335 b 797 365 828 354 817 365 l 738 365 l 738 485 l 867 485 b 956 396 929 485 956 458 l 956 0 l 833 0 l 833 61 b 835 103 833 82 835 103 l 832 103 b 526 -17 832 103 722 -17 b 49 494 265 -17 49 189 m 476 1225 l 600 1225 l 722 1049 l 613 1049 l 540 1160 l 538 1160 l 464 1049 l 356 1049 "},"ĝ":{"ha":826,"x_min":0,"x_max":0,"o":"m 51 368 b 360 719 51 569 164 719 b 585 610 535 719 585 610 l 588 610 b 588 632 588 610 588 621 b 665 703 588 674 610 703 l 796 703 l 796 588 l 740 588 b 710 560 722 588 710 578 l 710 29 b 356 -292 710 -203 533 -292 b 115 -233 271 -292 185 -268 l 161 -125 b 353 -172 161 -125 247 -172 b 575 24 475 -172 575 -118 l 575 71 b 578 115 575 93 578 115 l 575 115 b 365 8 533 47 468 8 b 51 368 171 8 51 165 m 392 125 b 579 367 489 125 579 183 b 381 603 579 550 489 603 b 189 371 258 603 189 514 b 392 125 189 225 265 125 m 356 981 l 479 981 l 601 804 l 492 804 l 419 915 l 417 915 l 343 804 l 235 804 "},"Ğ":{"ha":1022,"x_min":0,"x_max":0,"o":"m 49 494 b 543 997 49 776 263 997 b 893 876 778 997 893 876 l 814 775 b 546 869 814 775 708 869 b 193 494 350 869 193 713 b 549 114 193 267 350 114 b 828 236 724 114 828 236 l 828 335 b 797 365 828 354 817 365 l 738 365 l 738 485 l 867 485 b 956 396 929 485 956 458 l 956 0 l 833 0 l 833 61 b 835 103 833 82 835 103 l 832 103 b 526 -17 832 103 722 -17 b 49 494 265 -17 49 189 m 539 1043 b 361 1206 451 1043 361 1099 l 361 1225 l 454 1225 l 454 1207 b 539 1124 454 1149 496 1124 b 625 1208 582 1124 625 1149 l 625 1225 l 717 1225 l 717 1206 b 539 1043 717 1099 628 1043 "},"ğ":{"ha":826,"x_min":0,"x_max":0,"o":"m 51 368 b 360 719 51 569 164 719 b 585 610 535 719 585 610 l 588 610 b 588 632 588 610 588 621 b 665 703 588 674 610 703 l 796 703 l 796 588 l 740 588 b 710 560 722 588 710 578 l 710 29 b 356 -292 710 -203 533 -292 b 115 -233 271 -292 185 -268 l 161 -125 b 353 -172 161 -125 247 -172 b 575 24 475 -172 575 -118 l 575 71 b 578 115 575 93 578 115 l 575 115 b 365 8 533 47 468 8 b 51 368 171 8 51 165 m 392 125 b 579 367 489 125 579 183 b 381 603 579 550 489 603 b 189 371 258 603 189 514 b 392 125 189 225 265 125 m 417 799 b 239 961 329 799 239 854 l 239 981 l 332 981 l 332 963 b 417 879 332 904 374 879 b 503 964 460 879 503 904 l 503 981 l 594 981 l 594 961 b 417 799 594 854 506 799 "},"Ġ":{"ha":1022,"x_min":0,"x_max":0,"o":"m 49 494 b 543 997 49 776 263 997 b 893 876 778 997 893 876 l 814 775 b 546 869 814 775 708 869 b 193 494 350 869 193 713 b 549 114 193 267 350 114 b 828 236 724 114 828 236 l 828 335 b 797 365 828 354 817 365 l 738 365 l 738 485 l 867 485 b 956 396 929 485 956 458 l 956 0 l 833 0 l 833 61 b 835 103 833 82 835 103 l 832 103 b 526 -17 832 103 722 -17 b 49 494 265 -17 49 189 m 478 1225 l 599 1225 l 599 1085 l 478 1085 "},"ġ":{"ha":826,"x_min":0,"x_max":0,"o":"m 51 368 b 360 719 51 569 164 719 b 585 610 535 719 585 610 l 588 610 b 588 632 588 610 588 621 b 665 703 588 674 610 703 l 796 703 l 796 588 l 740 588 b 710 560 722 588 710 578 l 710 29 b 356 -292 710 -203 533 -292 b 115 -233 271 -292 185 -268 l 161 -125 b 353 -172 161 -125 247 -172 b 575 24 475 -172 575 -118 l 575 71 b 578 115 575 93 578 115 l 575 115 b 365 8 533 47 468 8 b 51 368 171 8 51 165 m 392 125 b 579 367 489 125 579 183 b 381 603 579 550 489 603 b 189 371 258 603 189 514 b 392 125 189 225 265 125 m 357 981 l 478 981 l 478 840 l 357 840 "},"Ģ":{"ha":1022,"x_min":0,"x_max":0,"o":"m 49 494 b 543 997 49 776 263 997 b 893 876 778 997 893 876 l 814 775 b 546 869 814 775 708 869 b 193 494 350 869 193 713 b 549 114 193 267 350 114 b 828 236 724 114 828 236 l 828 335 b 797 365 828 354 817 365 l 738 365 l 738 485 l 867 485 b 956 396 929 485 956 458 l 956 0 l 833 0 l 833 61 b 835 103 833 82 835 103 l 832 103 b 526 -17 832 103 722 -17 b 49 494 265 -17 49 189 m 583 -82 b 592 -150 589 -107 592 -129 b 558 -278 592 -193 581 -231 l 463 -278 b 489 -164 479 -235 489 -200 b 474 -82 489 -139 485 -112 "},"ģ":{"ha":826,"x_min":0,"x_max":0,"o":"m 51 368 b 360 719 51 569 164 719 b 585 610 535 719 585 610 l 588 610 b 588 632 588 610 588 621 b 665 703 588 674 610 703 l 796 703 l 796 588 l 740 588 b 710 560 722 588 710 578 l 710 29 b 356 -292 710 -203 533 -292 b 115 -233 271 -292 185 -268 l 161 -125 b 353 -172 161 -125 247 -172 b 575 24 475 -172 575 -118 l 575 71 b 578 115 575 93 578 115 l 575 115 b 365 8 533 47 468 8 b 51 368 171 8 51 165 m 392 125 b 579 367 489 125 579 183 b 381 603 579 550 489 603 b 189 371 258 603 189 514 b 392 125 189 225 265 125 m 344 785 b 336 853 339 810 336 832 b 369 981 336 896 347 933 l 465 981 b 439 867 449 938 439 903 b 454 785 439 842 443 815 "},"Ĥ":{"ha":1063,"x_min":0,"x_max":0,"o":"m 143 831 b 113 861 143 851 132 861 l 53 861 l 53 981 l 193 981 b 282 892 256 981 282 954 l 282 549 l 781 549 l 781 892 b 869 981 781 954 807 981 l 1010 981 l 1010 861 l 950 861 b 919 831 931 861 919 851 l 919 0 l 781 0 l 781 431 l 282 431 l 282 0 l 143 0 m 469 1225 l 593 1225 l 715 1049 l 606 1049 l 533 1160 l 531 1160 l 457 1049 l 349 1049 "},"ĥ":{"ha":864,"x_min":0,"x_max":0,"o":"m 113 835 b 82 865 113 856 101 865 l 25 865 l 25 981 l 158 981 b 247 894 221 981 247 954 l 247 621 b 244 563 247 588 244 563 l 247 563 b 508 719 278 631 368 719 b 749 450 674 719 749 629 l 749 146 b 779 115 749 125 760 115 l 836 115 l 836 0 l 704 0 b 615 89 642 0 615 26 l 615 419 b 482 596 615 518 594 596 b 260 422 375 596 289 524 b 247 329 250 394 247 363 l 247 0 l 113 0 m 386 1225 l 510 1225 l 632 1049 l 522 1049 l 450 1160 l 447 1160 l 374 1049 l 265 1049 "},"Ħ":{"ha":1063,"x_min":0,"x_max":0,"o":"m 143 714 l 51 714 l 51 808 l 143 808 l 143 981 l 282 981 l 282 808 l 781 808 l 781 981 l 919 981 l 919 808 l 1010 808 l 1010 714 l 919 714 l 919 0 l 781 0 l 781 431 l 282 431 l 282 0 l 143 0 m 781 549 l 781 714 l 282 714 l 282 549 "},"ħ":{"ha":864,"x_min":0,"x_max":0,"o":"m 113 782 l 7 782 l 7 869 l 113 869 l 113 981 l 247 981 l 247 869 l 426 869 l 426 782 l 247 782 l 247 621 b 244 563 247 588 244 563 l 247 563 b 508 719 261 600 340 719 b 749 450 674 719 749 629 l 749 146 b 779 115 749 125 760 115 l 836 115 l 836 0 l 704 0 b 615 89 642 0 615 26 l 615 419 b 482 596 615 515 594 596 b 247 329 340 596 247 478 l 247 0 l 113 0 "},"Ĩ":{"ha":440,"x_min":0,"x_max":0,"o":"m 58 119 l 154 119 l 154 861 l 58 861 l 58 981 l 382 981 l 382 861 l 286 861 l 286 119 l 382 119 l 382 0 l 58 0 m -17 1050 b 126 1226 -17 1167 32 1226 b 304 1132 222 1226 243 1132 b 360 1222 346 1132 360 1182 l 451 1222 b 308 1046 451 1107 403 1046 b 131 1140 214 1046 192 1140 b 75 1050 89 1140 75 1092 l -17 1050 "},"ĩ":{"ha":385,"x_min":0,"x_max":0,"o":"m 125 557 b 94 588 125 578 114 588 l 38 588 l 38 703 l 171 703 b 260 614 233 703 260 676 l 260 146 b 290 115 260 125 271 115 l 347 115 l 347 0 l 215 0 b 125 89 151 0 125 26 m -24 807 b 104 983 -24 922 14 983 b 249 892 183 983 207 892 b 294 979 286 892 294 938 l 385 979 b 257 803 385 863 347 803 b 113 893 176 803 153 893 b 67 807 75 893 67 847 l -24 807 "},"Ī":{"ha":440,"x_min":0,"x_max":0,"o":"m 58 119 l 154 119 l 154 861 l 58 861 l 58 981 l 382 981 l 382 861 l 286 861 l 286 119 l 382 119 l 382 0 l 58 0 m 43 1183 l 396 1183 l 396 1089 l 43 1089 "},"ī":{"ha":385,"x_min":0,"x_max":0,"o":"m 125 557 b 94 588 125 578 114 588 l 38 588 l 38 703 l 171 703 b 260 614 233 703 260 676 l 260 146 b 290 115 260 125 271 115 l 347 115 l 347 0 l 215 0 b 125 89 151 0 125 26 m 22 940 l 343 940 l 343 846 l 22 846 "},"Ĭ":{"ha":440,"x_min":0,"x_max":0,"o":"m 58 119 l 154 119 l 154 861 l 58 861 l 58 981 l 382 981 l 382 861 l 286 861 l 286 119 l 382 119 l 382 0 l 58 0 m 219 1043 b 42 1206 132 1043 42 1099 l 42 1225 l 135 1225 l 135 1207 b 219 1124 135 1149 176 1124 b 306 1208 263 1124 306 1149 l 306 1225 l 397 1225 l 397 1206 b 219 1043 397 1099 308 1043 "},"ĭ":{"ha":385,"x_min":0,"x_max":0,"o":"m 125 557 b 94 588 125 578 114 588 l 38 588 l 38 703 l 171 703 b 260 614 233 703 260 676 l 260 146 b 290 115 260 125 271 115 l 347 115 l 347 0 l 215 0 b 125 89 151 0 125 26 m 182 799 b 4 963 94 799 4 854 l 4 981 l 97 981 l 97 964 b 182 879 97 904 139 879 b 268 964 225 879 268 904 l 268 981 l 360 981 l 360 963 b 182 799 360 854 271 799 "},"Į":{"ha":440,"x_min":0,"x_max":0,"o":"m 58 119 l 154 119 l 154 861 l 58 861 l 58 981 l 382 981 l 382 861 l 286 861 l 286 119 l 382 119 l 382 0 b 293 -124 354 -24 293 -75 b 369 -183 293 -176 369 -183 l 339 -287 b 176 -156 339 -287 176 -275 b 267 -3 176 -86 226 -35 l 267 0 l 58 0 "},"į":{"ha":382,"x_min":0,"x_max":0,"o":"m 128 981 l 249 981 l 249 840 l 128 840 m 142 -156 b 231 -3 142 -86 190 -33 l 231 0 l 203 0 b 125 113 144 0 125 42 l 125 557 b 94 588 125 578 114 588 l 38 588 l 38 703 l 132 703 b 258 567 244 703 258 660 l 258 146 b 289 115 258 125 269 115 l 346 115 l 346 0 b 257 -124 318 -24 257 -75 b 335 -183 257 -176 335 -183 l 303 -287 b 142 -156 303 -287 142 -275 "},"İ":{"ha":440,"x_min":0,"x_max":0,"o":"m 58 119 l 154 119 l 154 861 l 58 861 l 58 981 l 382 981 l 382 861 l 286 861 l 286 119 l 382 119 l 382 0 l 58 0 m 158 1225 l 279 1225 l 279 1085 l 158 1085 "},"ı":{"ha":385,"x_min":0,"x_max":0,"o":"m 125 557 b 94 588 125 578 114 588 l 38 588 l 38 703 l 171 703 b 260 614 233 703 260 676 l 260 146 b 290 115 260 125 271 115 l 347 115 l 347 0 l 215 0 b 125 89 151 0 125 26 "},"Ĳ":{"ha":946,"x_min":0,"x_max":0,"o":"m 506 119 b 697 333 585 119 697 151 l 697 831 b 667 861 697 851 686 861 l 511 861 l 511 981 l 747 981 b 836 892 810 981 836 954 l 836 332 b 506 -7 836 49 656 -7 m 58 119 l 154 119 l 154 861 l 58 861 l 58 981 l 382 981 l 382 861 l 286 861 l 286 119 l 382 119 l 382 0 l 58 0 "},"ĳ":{"ha":765,"x_min":0,"x_max":0,"o":"m 128 981 l 249 981 l 249 840 l 128 840 m 125 557 b 94 588 125 578 114 588 l 38 588 l 38 703 l 169 703 b 258 614 232 703 258 676 l 258 146 b 289 115 258 125 269 115 l 346 115 l 346 0 l 214 0 b 125 89 151 0 125 26 m 524 981 l 644 981 l 644 840 l 524 840 m 340 -168 b 371 -169 340 -168 353 -169 b 521 -15 422 -169 521 -153 l 521 557 b 490 588 521 578 510 588 l 433 588 l 433 703 l 567 703 b 656 614 629 703 656 676 l 656 -24 b 386 -286 656 -258 474 -286 b 340 -283 358 -286 340 -283 l 340 -168 "},"Ĵ":{"ha":725,"x_min":0,"x_max":0,"o":"m 31 331 l 169 331 l 169 293 b 324 114 169 157 247 114 b 476 288 401 114 476 160 l 476 831 b 446 861 476 851 465 861 l 206 861 l 206 981 l 526 981 b 615 892 589 981 615 954 l 615 282 b 322 -17 615 69 467 -17 b 31 282 178 -17 31 71 m 339 1225 l 463 1225 l 585 1049 l 475 1049 l 403 1160 l 400 1160 l 326 1049 l 218 1049 "},"ĵ":{"ha":394,"x_min":0,"x_max":0,"o":"m 121 981 l 244 981 l 365 804 l 257 804 l 182 918 l 108 804 l -1 804 m -28 -171 b 3 -172 -28 -171 -15 -172 b 153 -17 54 -172 153 -154 l 153 561 b 118 592 153 585 142 592 l 65 592 l 65 703 l 196 703 b 285 622 258 703 285 685 l 285 -24 b 18 -286 285 -258 107 -286 b -28 -283 -10 -286 -28 -283 l -28 -171 "},"Ķ":{"ha":849,"x_min":0,"x_max":0,"o":"m 143 831 b 113 861 143 851 132 861 l 53 861 l 53 981 l 193 981 b 282 892 256 981 282 954 l 282 568 l 375 568 b 456 601 411 568 439 574 l 674 981 l 829 981 l 592 574 b 536 510 564 525 536 510 l 536 507 b 592 446 536 507 564 500 l 739 151 b 808 119 753 122 775 119 l 835 119 l 835 0 l 758 0 b 622 74 682 0 654 10 l 453 414 b 365 450 436 446 406 450 l 282 450 l 282 0 l 143 0 m 506 -82 b 514 -150 511 -107 514 -129 b 481 -278 514 -193 503 -231 l 385 -278 b 411 -164 401 -235 411 -200 b 396 -82 411 -139 407 -112 "},"ķ":{"ha":729,"x_min":0,"x_max":0,"o":"m 113 835 b 82 865 113 856 101 865 l 25 865 l 25 981 l 158 981 b 247 900 221 981 247 963 l 247 444 l 310 444 b 383 475 335 444 365 451 l 546 703 l 700 703 l 510 442 b 457 389 478 399 457 389 l 457 386 b 504 333 457 386 481 378 l 608 143 b 668 115 619 121 633 115 l 711 115 l 711 0 l 614 0 b 497 65 549 0 528 11 l 369 301 b 300 331 354 328 322 331 l 247 331 l 247 0 l 113 0 m 442 -82 b 450 -150 447 -107 450 -129 b 417 -278 450 -193 439 -231 l 321 -278 b 347 -164 338 -235 347 -200 b 332 -82 347 -139 343 -112 "},"ĸ":{"ha":729,"x_min":0,"x_max":0,"o":"m 113 557 b 82 588 113 578 101 588 l 25 588 l 25 703 l 158 703 b 247 622 221 703 247 685 l 247 417 l 310 417 b 383 446 335 417 365 419 l 554 703 l 707 703 l 508 415 b 458 364 478 371 458 364 l 458 361 b 503 307 458 361 478 353 l 593 143 b 654 115 604 121 618 115 l 703 115 l 703 0 l 599 0 b 482 65 533 0 513 11 l 369 274 b 300 301 354 301 322 301 l 247 301 l 247 0 l 113 0 "},"Ĺ":{"ha":754,"x_min":0,"x_max":0,"o":"m 143 831 b 113 861 143 851 132 861 l 53 861 l 53 981 l 193 981 b 282 892 256 981 282 954 l 282 150 b 313 119 282 131 293 119 l 583 119 b 614 150 603 119 614 131 l 614 213 l 740 213 l 740 89 b 651 0 740 26 714 0 l 232 0 b 143 89 169 0 143 26 m 271 1225 l 407 1225 l 275 1049 l 169 1049 "},"ĺ":{"ha":372,"x_min":0,"x_max":0,"o":"m 115 835 b 85 865 115 856 104 865 l 28 865 l 28 981 l 161 981 b 250 892 224 981 250 954 l 250 146 b 281 115 250 125 261 115 l 338 115 l 338 0 l 204 0 b 115 89 142 0 115 26 m 203 1225 l 339 1225 l 207 1049 l 101 1049 "},"Ļ":{"ha":754,"x_min":0,"x_max":0,"o":"m 143 831 b 113 861 143 851 132 861 l 53 861 l 53 981 l 193 981 b 282 892 256 981 282 954 l 282 150 b 313 119 282 131 293 119 l 583 119 b 614 150 603 119 614 131 l 614 213 l 740 213 l 740 89 b 651 0 740 26 714 0 l 232 0 b 143 89 169 0 143 26 m 492 -82 b 500 -150 497 -107 500 -129 b 467 -278 500 -193 489 -231 l 371 -278 b 397 -164 388 -235 397 -200 b 382 -82 397 -139 393 -112 "},"ļ":{"ha":372,"x_min":0,"x_max":0,"o":"m 115 835 b 85 865 115 856 104 865 l 28 865 l 28 981 l 161 981 b 250 892 224 981 250 954 l 250 146 b 281 115 250 125 261 115 l 338 115 l 338 0 l 204 0 b 115 89 142 0 115 26 m 254 -82 b 263 -150 260 -107 263 -129 b 229 -278 263 -193 251 -231 l 133 -278 b 160 -164 150 -235 160 -200 b 144 -82 160 -139 156 -112 "},"Ľ":{"ha":754,"x_min":0,"x_max":0,"o":"m 143 831 b 113 861 143 851 132 861 l 53 861 l 53 981 l 193 981 b 282 892 256 981 282 954 l 282 150 b 313 119 282 131 293 119 l 583 119 b 614 150 603 119 614 131 l 614 213 l 740 213 l 740 89 b 651 0 740 26 714 0 l 232 0 b 143 89 169 0 143 26 m 510 981 b 518 913 515 956 518 933 b 485 785 518 869 507 832 l 389 785 b 415 899 406 828 415 863 b 400 981 415 924 411 950 "},"ľ":{"ha":492,"x_min":0,"x_max":0,"o":"m 115 835 b 85 865 115 856 104 865 l 28 865 l 28 981 l 161 981 b 250 892 224 981 250 954 l 250 146 b 281 115 250 125 261 115 l 338 115 l 338 0 l 204 0 b 115 89 142 0 115 26 m 438 981 b 446 913 443 956 446 933 b 413 785 446 869 435 832 l 317 785 b 343 899 333 828 343 863 b 328 981 343 924 339 950 "},"Ŀ":{"ha":754,"x_min":0,"x_max":0,"o":"m 143 831 b 113 861 143 851 132 861 l 53 861 l 53 981 l 193 981 b 282 892 256 981 282 954 l 282 150 b 313 119 282 131 293 119 l 583 119 b 614 150 603 119 614 131 l 614 213 l 740 213 l 740 89 b 651 0 740 26 714 0 l 232 0 b 143 89 169 0 143 26 m 438 599 l 575 599 l 575 456 l 438 456 "},"ŀ":{"ha":494,"x_min":0,"x_max":0,"o":"m 115 835 b 85 865 115 856 104 865 l 28 865 l 28 981 l 161 981 b 250 892 224 981 250 954 l 250 146 b 281 115 250 125 261 115 l 338 115 l 338 0 l 204 0 b 115 89 142 0 115 26 m 349 599 l 486 599 l 486 456 l 349 456 "},"Ł":{"ha":754,"x_min":0,"x_max":0,"o":"m 143 421 l 56 371 l 56 506 l 143 556 l 143 831 b 113 861 143 851 132 861 l 53 861 l 53 981 l 193 981 b 282 892 256 981 282 954 l 282 635 l 426 718 l 426 583 l 282 500 l 282 150 b 313 119 282 131 293 119 l 583 119 b 614 150 603 119 614 131 l 614 213 l 740 213 l 740 89 b 651 0 740 26 714 0 l 232 0 b 143 89 169 0 143 26 "},"ł":{"ha":396,"x_min":0,"x_max":0,"o":"m 131 446 l 26 386 l 26 521 l 131 582 l 131 835 b 100 865 131 856 119 865 l 44 865 l 44 981 l 176 981 b 265 892 239 981 265 954 l 265 658 l 374 721 l 374 586 l 265 522 l 265 146 b 296 115 265 125 276 115 l 353 115 l 353 0 l 221 0 b 131 89 158 0 131 26 "},"Ń":{"ha":1063,"x_min":0,"x_max":0,"o":"m 53 119 l 113 119 b 143 150 132 119 143 131 l 143 981 l 269 981 l 706 360 b 788 226 742 307 788 226 l 790 226 b 781 360 790 226 781 304 l 781 892 b 871 981 781 954 807 981 l 1010 981 l 1010 861 l 950 861 b 919 831 931 861 919 851 l 919 0 l 794 0 l 357 621 b 276 754 321 674 276 754 l 274 754 b 282 621 274 754 282 675 l 282 89 b 193 0 282 26 256 0 l 53 0 m 592 1225 l 728 1225 l 596 1049 l 490 1049 "},"ń":{"ha":869,"x_min":0,"x_max":0,"o":"m 118 557 b 88 588 118 578 107 588 l 31 588 l 31 703 l 160 703 b 249 625 219 703 249 676 l 249 599 b 246 558 249 576 246 558 l 249 558 b 513 719 276 619 360 719 b 754 450 679 719 754 629 l 754 146 b 785 115 754 125 765 115 l 842 115 l 842 0 l 708 0 b 619 89 646 0 619 26 l 619 419 b 488 596 619 518 600 596 b 265 424 379 596 293 525 b 253 329 256 396 253 364 l 253 0 l 118 0 m 496 981 l 632 981 l 500 804 l 394 804 "},"Ņ":{"ha":1063,"x_min":0,"x_max":0,"o":"m 53 119 l 113 119 b 143 150 132 119 143 131 l 143 981 l 269 981 l 706 360 b 788 226 742 307 788 226 l 790 226 b 781 360 790 226 781 304 l 781 892 b 871 981 781 954 807 981 l 1010 981 l 1010 861 l 950 861 b 919 831 931 861 919 851 l 919 0 l 794 0 l 357 621 b 276 754 321 674 276 754 l 274 754 b 282 621 274 754 282 675 l 282 89 b 193 0 282 26 256 0 l 53 0 m 581 -82 b 589 -150 586 -107 589 -129 b 556 -278 589 -193 578 -231 l 460 -278 b 486 -164 476 -235 486 -200 b 471 -82 486 -139 482 -112 "},"ņ":{"ha":869,"x_min":0,"x_max":0,"o":"m 118 557 b 88 588 118 578 107 588 l 31 588 l 31 703 l 160 703 b 249 625 219 703 249 676 l 249 599 b 246 558 249 576 246 558 l 249 558 b 513 719 276 619 360 719 b 754 450 679 719 754 629 l 754 146 b 785 115 754 125 765 115 l 842 115 l 842 0 l 708 0 b 619 89 646 0 619 26 l 619 419 b 488 596 619 518 600 596 b 265 424 379 596 293 525 b 253 329 256 396 253 364 l 253 0 l 118 0 m 504 -82 b 513 -150 510 -107 513 -129 b 479 -278 513 -193 501 -231 l 383 -278 b 410 -164 400 -235 410 -200 b 394 -82 410 -139 406 -112 "},"Ň":{"ha":1063,"x_min":0,"x_max":0,"o":"m 53 119 l 113 119 b 143 150 132 119 143 131 l 143 981 l 269 981 l 706 360 b 788 226 742 307 788 226 l 790 226 b 781 360 790 226 781 304 l 781 892 b 871 981 781 954 807 981 l 1010 981 l 1010 861 l 950 861 b 919 831 931 861 919 851 l 919 0 l 794 0 l 357 621 b 276 754 321 674 276 754 l 274 754 b 282 621 274 754 282 675 l 282 89 b 193 0 282 26 256 0 l 53 0 m 361 1225 l 471 1225 l 543 1113 l 546 1113 l 618 1225 l 729 1225 l 607 1049 l 483 1049 "},"ň":{"ha":869,"x_min":0,"x_max":0,"o":"m 118 557 b 88 588 118 578 107 588 l 31 588 l 31 703 l 160 703 b 249 625 219 703 249 676 l 249 599 b 246 558 249 576 246 558 l 249 558 b 513 719 276 619 360 719 b 754 450 679 719 754 629 l 754 146 b 785 115 754 125 765 115 l 842 115 l 842 0 l 708 0 b 619 89 646 0 619 26 l 619 419 b 488 596 619 518 600 596 b 265 424 379 596 293 525 b 253 329 256 396 253 364 l 253 0 l 118 0 m 264 981 l 374 981 l 446 868 l 449 868 l 521 981 l 632 981 l 510 804 l 386 804 "},"ŉ":{"ha":908,"x_min":0,"x_max":0,"o":"m 157 557 b 126 588 157 578 146 588 l 69 588 l 69 703 l 199 703 b 288 625 258 703 288 676 l 288 599 b 285 558 288 576 285 558 l 288 558 b 551 719 315 619 399 719 b 793 450 718 719 793 629 l 793 146 b 824 115 793 125 804 115 l 881 115 l 881 0 l 747 0 b 658 89 685 0 658 26 l 658 419 b 526 596 658 518 639 596 b 304 424 418 596 332 525 b 292 329 294 396 292 364 l 292 0 l 157 0 m 101 981 b 110 913 107 956 110 933 b 76 785 110 869 99 832 l -19 785 b 7 899 -3 828 7 863 b -8 981 7 924 3 950 "},"Ŋ":{"ha":1063,"x_min":0,"x_max":0,"o":"m 53 119 l 113 119 b 143 150 132 119 143 131 l 143 981 l 269 981 l 706 360 b 788 226 742 307 788 226 l 790 226 b 781 360 790 226 781 304 l 781 892 b 871 981 781 954 807 981 l 1010 981 l 1010 861 l 950 861 b 919 831 931 861 919 851 l 919 38 b 597 -285 919 -215 747 -285 l 597 -158 b 781 13 672 -158 781 -114 l 781 22 l 357 621 b 276 754 321 674 276 754 l 274 754 b 282 621 274 754 282 675 l 282 89 b 193 0 282 26 256 0 l 53 0 "},"ŋ":{"ha":863,"x_min":0,"x_max":0,"o":"m 118 557 b 88 588 118 578 107 588 l 31 588 l 31 703 l 160 703 b 249 625 219 703 249 676 l 249 599 b 246 558 249 576 246 558 l 249 558 b 513 719 276 619 360 719 b 754 450 679 719 754 629 l 754 -28 b 485 -290 754 -262 572 -290 b 438 -287 457 -290 438 -287 l 438 -172 b 468 -174 438 -172 450 -174 b 619 -18 521 -174 619 -157 l 619 419 b 488 596 619 518 600 596 b 264 422 381 596 293 524 b 253 329 256 394 253 363 l 253 0 l 118 0 "},"Ō":{"ha":1101,"x_min":0,"x_max":0,"o":"m 47 497 b 550 997 47 779 267 997 b 1054 497 833 997 1054 779 b 550 -17 1054 208 833 -17 b 47 497 267 -17 47 208 m 192 497 b 550 114 192 281 351 114 b 910 497 749 114 910 281 b 550 868 910 707 749 868 b 192 497 351 868 192 707 m 371 1183 l 724 1183 l 724 1089 l 371 1089 "},"ō":{"ha":857,"x_min":0,"x_max":0,"o":"m 50 353 b 428 719 50 565 218 719 b 807 353 638 719 807 565 b 428 -17 807 139 638 -17 b 50 353 218 -17 50 139 m 188 353 b 428 100 188 206 296 100 b 669 353 561 100 669 206 b 428 603 669 499 561 603 b 188 353 296 603 188 499 m 253 939 l 606 939 l 606 844 l 253 844 "},"Ŏ":{"ha":1101,"x_min":0,"x_max":0,"o":"m 47 497 b 550 997 47 779 267 997 b 1054 497 833 997 1054 779 b 550 -17 1054 208 833 -17 b 47 497 267 -17 47 208 m 192 497 b 550 114 192 281 351 114 b 910 497 749 114 910 281 b 550 868 910 707 749 868 b 192 497 351 868 192 707 m 547 1043 b 369 1206 460 1043 369 1099 l 369 1225 l 463 1225 l 463 1207 b 547 1124 463 1149 504 1124 b 633 1208 590 1124 633 1149 l 633 1225 l 725 1225 l 725 1206 b 547 1043 725 1099 636 1043 "},"ŏ":{"ha":857,"x_min":0,"x_max":0,"o":"m 50 353 b 428 719 50 565 218 719 b 807 353 638 719 807 565 b 428 -17 807 139 638 -17 b 50 353 218 -17 50 139 m 188 353 b 428 100 188 206 296 100 b 669 353 561 100 669 206 b 428 603 669 499 561 603 b 188 353 296 603 188 499 m 428 799 b 250 961 340 799 250 854 l 250 981 l 343 981 l 343 963 b 428 879 343 904 385 879 b 514 964 471 879 514 904 l 514 981 l 606 981 l 606 961 b 428 799 606 854 517 799 "},"Ő":{"ha":1101,"x_min":0,"x_max":0,"o":"m 47 497 b 550 997 47 779 267 997 b 1054 497 833 997 1054 779 b 550 -17 1054 208 833 -17 b 47 497 267 -17 47 208 m 192 497 b 550 114 192 281 351 114 b 910 497 749 114 910 281 b 550 868 910 707 749 868 b 192 497 351 868 192 707 m 694 1225 l 832 1225 l 699 1049 l 594 1049 m 478 1225 l 594 1225 l 497 1049 l 382 1049 "},"ő":{"ha":857,"x_min":0,"x_max":0,"o":"m 50 353 b 428 719 50 565 218 719 b 807 353 638 719 807 565 b 428 -17 807 139 638 -17 b 50 353 218 -17 50 139 m 188 353 b 428 100 188 206 296 100 b 669 353 561 100 669 206 b 428 603 669 499 561 603 b 188 353 296 603 188 499 m 576 981 l 714 981 l 581 804 l 476 804 m 360 981 l 476 981 l 379 804 l 264 804 "},"Œ":{"ha":1290,"x_min":0,"x_max":0,"o":"m 47 492 b 529 993 47 774 254 993 b 729 981 606 993 672 981 l 1131 981 b 1221 892 1194 981 1221 954 l 1221 768 l 1093 768 l 1093 831 b 1063 861 1093 851 1082 861 l 792 861 l 792 554 l 1128 554 l 1128 435 l 792 435 l 792 119 l 1101 119 b 1132 150 1121 119 1132 131 l 1132 213 l 1258 213 l 1258 89 b 1174 0 1258 26 1232 0 l 729 0 b 529 -12 674 0 606 -12 b 47 492 254 -12 47 208 m 192 492 b 529 117 192 278 329 117 b 653 128 578 117 624 121 l 653 853 b 529 864 622 860 578 864 b 192 492 329 864 192 704 "},"œ":{"ha":1392,"x_min":0,"x_max":0,"o":"m 50 350 b 426 719 50 572 221 719 b 735 553 568 719 674 654 l 738 553 b 1026 719 793 657 892 719 b 1331 390 1225 719 1331 572 b 1326 335 1331 372 1326 335 l 814 335 b 1056 100 825 181 933 100 b 1261 182 1176 100 1261 182 l 1319 86 b 1047 -17 1319 86 1215 -17 b 740 154 901 -17 797 51 l 738 154 b 428 -17 676 50 569 -17 b 50 350 221 -17 50 128 m 1192 433 b 1026 613 1186 550 1117 613 b 819 433 922 613 840 550 m 188 350 b 428 100 188 197 296 100 b 669 354 558 100 669 197 b 428 603 669 506 556 603 b 188 350 296 603 188 503 "},"Ŕ":{"ha":858,"x_min":0,"x_max":0,"o":"m 143 861 l 53 861 l 53 981 l 481 981 b 774 693 653 981 774 872 b 597 433 774 529 667 449 l 597 431 b 639 386 597 431 622 418 l 757 151 b 826 119 771 122 793 119 l 844 119 l 844 0 l 776 0 b 640 74 700 0 672 10 l 507 338 b 410 390 482 383 464 390 l 282 390 l 282 0 l 143 0 m 463 510 b 631 688 568 510 631 578 b 465 861 631 797 568 861 l 282 861 l 282 510 m 446 1225 l 582 1225 l 450 1049 l 344 1049 "},"ŕ":{"ha":533,"x_min":0,"x_max":0,"o":"m 118 557 b 88 588 118 578 107 588 l 31 588 l 31 703 l 158 703 b 249 621 219 703 249 678 l 249 576 b 246 528 249 549 246 528 l 249 528 b 471 711 282 631 363 711 b 507 708 489 711 507 708 l 507 575 b 467 579 507 575 489 579 b 269 414 381 579 301 518 b 253 288 257 374 253 331 l 253 0 l 118 0 m 324 981 l 460 981 l 328 804 l 222 804 "},"Ŗ":{"ha":858,"x_min":0,"x_max":0,"o":"m 143 861 l 53 861 l 53 981 l 481 981 b 774 693 653 981 774 872 b 597 433 774 529 667 449 l 597 431 b 639 386 597 431 622 418 l 757 151 b 826 119 771 122 793 119 l 844 119 l 844 0 l 776 0 b 640 74 700 0 672 10 l 507 338 b 410 390 482 383 464 390 l 282 390 l 282 0 l 143 0 m 463 510 b 631 688 568 510 631 578 b 465 861 631 797 568 861 l 282 861 l 282 510 m 528 -82 b 536 -150 533 -107 536 -129 b 503 -278 536 -193 525 -231 l 407 -278 b 433 -164 424 -235 433 -200 b 418 -82 433 -139 429 -112 "},"ŗ":{"ha":533,"x_min":0,"x_max":0,"o":"m 118 557 b 88 588 118 578 107 588 l 31 588 l 31 703 l 158 703 b 249 621 219 703 249 678 l 249 576 b 246 528 249 549 246 528 l 249 528 b 471 711 282 631 363 711 b 507 708 489 711 507 708 l 507 575 b 467 579 507 575 489 579 b 269 414 381 579 301 518 b 253 288 257 374 253 331 l 253 0 l 118 0 m 246 -82 b 254 -150 251 -107 254 -129 b 221 -278 254 -193 243 -231 l 125 -278 b 151 -164 142 -235 151 -200 b 136 -82 151 -139 147 -112 "},"Ř":{"ha":858,"x_min":0,"x_max":0,"o":"m 143 861 l 53 861 l 53 981 l 481 981 b 774 693 653 981 774 872 b 597 433 774 529 667 449 l 597 431 b 639 386 597 431 622 418 l 757 151 b 826 119 771 122 793 119 l 844 119 l 844 0 l 776 0 b 640 74 700 0 672 10 l 507 338 b 410 390 482 383 464 390 l 282 390 l 282 0 l 143 0 m 463 510 b 631 688 568 510 631 578 b 465 861 631 797 568 861 l 282 861 l 282 510 m 215 1225 l 325 1225 l 397 1113 l 400 1113 l 472 1225 l 583 1225 l 461 1049 l 338 1049 "},"ř":{"ha":533,"x_min":0,"x_max":0,"o":"m 118 557 b 88 588 118 578 107 588 l 31 588 l 31 703 l 158 703 b 249 621 219 703 249 678 l 249 576 b 246 528 249 549 246 528 l 249 528 b 471 711 282 631 363 711 b 507 708 489 711 507 708 l 507 575 b 467 579 507 575 489 579 b 269 414 381 579 301 518 b 253 288 257 374 253 331 l 253 0 l 118 0 m 93 981 l 203 981 l 275 868 l 278 868 l 350 981 l 461 981 l 339 804 l 215 804 "},"Ś":{"ha":740,"x_min":0,"x_max":0,"o":"m 129 228 b 375 110 129 228 229 110 b 543 256 467 110 543 165 b 69 735 543 461 69 429 b 383 997 69 882 199 997 b 663 817 488 997 663 950 l 663 736 l 535 736 l 535 782 b 383 871 535 833 461 871 b 211 739 278 871 211 813 b 683 263 211 540 683 585 b 372 -17 683 113 567 -17 b 47 129 164 -17 47 129 m 426 1225 l 563 1225 l 431 1049 l 325 1049 "},"ś":{"ha":631,"x_min":0,"x_max":0,"o":"m 107 197 b 319 93 107 197 190 93 b 442 183 385 93 442 122 b 58 522 442 310 58 296 b 319 719 58 657 176 719 b 550 583 404 719 550 690 l 550 517 l 429 517 l 429 551 b 324 611 429 593 367 611 b 193 528 244 611 193 583 b 576 189 193 394 576 424 b 318 -17 576 64 465 -17 b 38 107 131 -17 38 107 m 360 981 l 496 981 l 364 804 l 258 804 "},"Ŝ":{"ha":740,"x_min":0,"x_max":0,"o":"m 129 228 b 375 110 129 228 229 110 b 543 256 467 110 543 165 b 69 735 543 461 69 429 b 383 997 69 882 199 997 b 663 817 488 997 663 950 l 663 736 l 535 736 l 535 782 b 383 871 535 833 461 871 b 211 739 278 871 211 813 b 683 263 211 540 683 585 b 372 -17 683 113 567 -17 b 47 129 164 -17 47 129 m 318 1225 l 442 1225 l 564 1049 l 454 1049 l 382 1160 l 379 1160 l 306 1049 l 197 1049 "},"ŝ":{"ha":631,"x_min":0,"x_max":0,"o":"m 107 197 b 319 93 107 197 190 93 b 442 183 385 93 442 122 b 58 522 442 310 58 296 b 319 719 58 657 176 719 b 550 583 404 719 550 690 l 550 517 l 429 517 l 429 551 b 324 611 429 593 367 611 b 193 528 244 611 193 583 b 576 189 193 394 576 424 b 318 -17 576 64 465 -17 b 38 107 131 -17 38 107 m 251 981 l 375 981 l 497 804 l 388 804 l 315 915 l 313 915 l 239 804 l 131 804 "},"Ş":{"ha":740,"x_min":0,"x_max":0,"o":"m 129 228 b 375 110 129 228 229 110 b 543 256 467 110 543 165 b 69 735 543 461 69 429 b 383 997 69 882 199 997 b 663 817 488 997 663 950 l 663 736 l 535 736 l 535 782 b 383 871 535 833 461 871 b 211 739 278 871 211 813 b 683 263 211 540 683 585 b 401 -15 683 119 578 -4 l 390 -69 b 490 -171 447 -76 490 -111 b 349 -287 490 -257 419 -287 b 271 -278 304 -287 271 -278 l 271 -197 b 331 -207 271 -197 300 -207 b 396 -164 365 -207 396 -196 b 324 -121 396 -135 367 -121 b 299 -122 311 -121 299 -122 l 325 -14 b 47 129 146 6 47 129 "},"ş":{"ha":631,"x_min":0,"x_max":0,"o":"m 107 197 b 319 93 107 197 190 93 b 442 183 385 93 442 122 b 58 522 442 310 58 296 b 319 719 58 657 176 719 b 550 583 404 719 550 690 l 550 517 l 429 517 l 429 551 b 324 611 429 593 367 611 b 193 528 244 611 193 583 b 576 189 193 394 576 424 b 335 -17 576 68 472 -10 l 324 -69 b 422 -171 381 -76 422 -111 b 282 -287 422 -257 353 -287 b 204 -278 238 -287 204 -278 l 204 -197 b 264 -207 204 -197 232 -207 b 329 -164 299 -207 329 -196 b 257 -121 329 -135 300 -121 b 232 -122 244 -121 232 -122 l 260 -12 b 38 107 111 8 38 107 "},"Š":{"ha":740,"x_min":0,"x_max":0,"o":"m 129 228 b 375 110 129 228 229 110 b 543 256 467 110 543 165 b 69 735 543 461 69 429 b 383 997 69 882 199 997 b 663 817 488 997 663 950 l 663 736 l 535 736 l 535 782 b 383 871 535 833 461 871 b 211 739 278 871 211 813 b 683 263 211 540 683 585 b 372 -17 683 113 567 -17 b 47 129 164 -17 47 129 m 196 1225 l 306 1225 l 378 1113 l 381 1113 l 453 1225 l 564 1225 l 442 1049 l 318 1049 "},"š":{"ha":631,"x_min":0,"x_max":0,"o":"m 107 197 b 319 93 107 197 190 93 b 442 183 385 93 442 122 b 58 522 442 310 58 296 b 319 719 58 657 176 719 b 550 583 404 719 550 690 l 550 517 l 429 517 l 429 551 b 324 611 429 593 367 611 b 193 528 244 611 193 583 b 576 189 193 394 576 424 b 318 -17 576 64 465 -17 b 38 107 131 -17 38 107 m 129 981 l 239 981 l 311 868 l 314 868 l 386 981 l 497 981 l 375 804 l 251 804 "},"Ţ":{"ha":883,"x_min":0,"x_max":0,"o":"m 511 861 l 511 0 l 478 0 l 464 -69 b 563 -171 519 -76 563 -111 b 421 -287 563 -256 493 -287 b 344 -278 378 -287 344 -278 l 344 -197 b 404 -207 344 -197 372 -207 b 469 -164 439 -207 469 -196 b 396 -121 469 -135 439 -121 b 372 -122 383 -121 372 -122 l 403 0 l 372 0 l 372 861 l 175 861 b 144 831 156 861 144 851 l 144 768 l 21 768 l 21 897 b 104 981 21 960 42 981 l 779 981 b 863 897 842 981 863 960 l 863 768 l 739 768 l 739 831 b 708 861 739 851 728 861 "},"ţ":{"ha":489,"x_min":0,"x_max":0,"o":"m 124 594 l 32 594 l 32 703 l 126 703 l 126 896 l 258 896 l 258 703 l 426 703 l 426 594 l 258 594 l 258 272 b 407 114 258 132 356 114 b 439 117 426 114 439 117 l 439 -3 b 392 -6 439 -3 421 -6 b 319 1 371 -6 346 -4 l 306 -69 b 404 -171 363 -76 404 -111 b 264 -287 404 -257 335 -287 b 186 -278 219 -287 186 -278 l 186 -197 b 246 -207 186 -197 214 -207 b 311 -164 281 -207 311 -196 b 239 -121 311 -135 282 -121 b 214 -122 226 -121 214 -122 l 251 22 b 124 257 183 54 124 119 "},"Ť":{"ha":883,"x_min":0,"x_max":0,"o":"m 372 861 l 175 861 b 144 831 156 861 144 851 l 144 768 l 21 768 l 21 897 b 104 981 21 960 42 981 l 779 981 b 863 897 842 981 863 960 l 863 768 l 739 768 l 739 831 b 708 861 739 851 728 861 l 511 861 l 511 0 l 372 0 m 258 1225 l 368 1225 l 440 1113 l 443 1113 l 515 1225 l 626 1225 l 504 1049 l 381 1049 "},"ť":{"ha":528,"x_min":0,"x_max":0,"o":"m 126 594 l 35 594 l 35 703 l 129 703 l 129 896 l 261 896 l 261 703 l 429 703 l 429 594 l 261 594 l 261 272 b 410 114 261 132 358 114 b 442 117 429 114 442 117 l 442 -3 b 394 -6 442 -3 424 -6 b 126 257 306 -6 126 22 m 458 981 b 467 913 464 956 467 933 b 433 785 467 869 456 832 l 338 785 b 364 899 354 828 364 863 b 349 981 364 924 360 950 "},"Ŧ":{"ha":883,"x_min":0,"x_max":0,"o":"m 372 431 l 158 431 l 158 536 l 372 536 l 372 861 l 175 861 b 144 831 156 861 144 851 l 144 768 l 21 768 l 21 897 b 104 981 21 960 42 981 l 779 981 b 863 897 842 981 863 960 l 863 768 l 739 768 l 739 831 b 708 861 739 851 728 861 l 511 861 l 511 536 l 735 536 l 735 431 l 511 431 l 511 0 l 372 0 "},"ŧ":{"ha":489,"x_min":0,"x_max":0,"o":"m 124 360 l 32 360 l 32 454 l 124 454 l 124 594 l 32 594 l 32 703 l 126 703 l 126 896 l 258 896 l 258 703 l 426 703 l 426 594 l 258 594 l 258 454 l 426 454 l 426 360 l 258 360 l 258 268 b 407 114 258 133 356 114 b 439 117 426 114 439 117 l 439 -3 b 392 -6 439 -3 421 -6 b 124 257 303 -6 124 22 "},"Ũ":{"ha":1011,"x_min":0,"x_max":0,"o":"m 132 831 b 101 861 132 851 121 861 l 42 861 l 42 981 l 182 981 b 271 892 244 981 271 954 l 271 349 b 506 114 271 201 364 114 b 740 350 647 114 740 201 l 740 892 b 831 981 740 954 768 981 l 969 981 l 969 861 l 910 861 b 879 831 890 861 879 851 l 879 343 b 507 -17 879 129 731 -17 b 132 343 282 -17 132 129 m 271 1050 b 414 1226 271 1167 319 1226 b 592 1132 510 1226 531 1132 b 647 1222 633 1132 647 1182 l 739 1222 b 596 1046 739 1107 690 1046 b 418 1140 501 1046 479 1140 b 363 1050 376 1140 363 1092 l 271 1050 "},"ũ":{"ha":863,"x_min":0,"x_max":0,"o":"m 115 557 b 85 588 115 578 104 588 l 28 588 l 28 703 l 160 703 b 250 617 224 703 250 676 l 250 283 b 381 108 250 186 269 108 b 610 374 525 108 610 235 l 610 703 l 744 703 l 744 146 b 775 115 744 125 756 115 l 832 115 l 832 0 l 703 0 b 614 78 643 0 614 28 l 614 104 b 615 144 614 126 615 144 l 613 144 b 354 -17 583 75 493 -17 b 115 253 196 -17 115 67 m 186 806 b 329 982 186 922 235 982 b 507 888 425 982 446 888 b 563 978 549 888 563 938 l 654 978 b 511 801 654 863 606 801 b 333 896 417 801 394 896 b 278 806 292 896 278 847 l 186 806 "},"Ū":{"ha":1011,"x_min":0,"x_max":0,"o":"m 132 831 b 101 861 132 851 121 861 l 42 861 l 42 981 l 182 981 b 271 892 244 981 271 954 l 271 349 b 506 114 271 201 364 114 b 740 350 647 114 740 201 l 740 892 b 831 981 740 954 768 981 l 969 981 l 969 861 l 910 861 b 879 831 890 861 879 851 l 879 343 b 507 -17 879 129 731 -17 b 132 343 282 -17 132 129 m 331 1183 l 683 1183 l 683 1089 l 331 1089 "},"ū":{"ha":863,"x_min":0,"x_max":0,"o":"m 115 557 b 85 588 115 578 104 588 l 28 588 l 28 703 l 160 703 b 250 617 224 703 250 676 l 250 283 b 381 108 250 186 269 108 b 610 374 525 108 610 235 l 610 703 l 744 703 l 744 146 b 775 115 744 125 756 115 l 832 115 l 832 0 l 703 0 b 614 78 643 0 614 28 l 614 104 b 615 144 614 126 615 144 l 613 144 b 354 -17 583 75 493 -17 b 115 253 196 -17 115 67 m 246 939 l 599 939 l 599 844 l 246 844 "},"Ŭ":{"ha":1011,"x_min":0,"x_max":0,"o":"m 132 831 b 101 861 132 851 121 861 l 42 861 l 42 981 l 182 981 b 271 892 244 981 271 954 l 271 349 b 506 114 271 201 364 114 b 740 350 647 114 740 201 l 740 892 b 831 981 740 954 768 981 l 969 981 l 969 861 l 910 861 b 879 831 890 861 879 851 l 879 343 b 507 -17 879 129 731 -17 b 132 343 282 -17 132 129 m 507 1043 b 329 1206 419 1043 329 1099 l 329 1225 l 422 1225 l 422 1207 b 507 1124 422 1149 464 1124 b 593 1208 550 1124 593 1149 l 593 1225 l 685 1225 l 685 1206 b 507 1043 685 1099 596 1043 "},"ŭ":{"ha":863,"x_min":0,"x_max":0,"o":"m 115 557 b 85 588 115 578 104 588 l 28 588 l 28 703 l 160 703 b 250 617 224 703 250 676 l 250 283 b 381 108 250 186 269 108 b 610 374 525 108 610 235 l 610 703 l 744 703 l 744 146 b 775 115 744 125 756 115 l 832 115 l 832 0 l 703 0 b 614 78 643 0 614 28 l 614 104 b 615 144 614 126 615 144 l 613 144 b 354 -17 583 75 493 -17 b 115 253 196 -17 115 67 m 422 799 b 244 961 335 799 244 854 l 244 981 l 338 981 l 338 963 b 422 879 338 904 379 879 b 508 964 465 879 508 904 l 508 981 l 600 981 l 600 961 b 422 799 600 854 511 799 "},"Ů":{"ha":1011,"x_min":0,"x_max":0,"o":"m 132 831 b 101 861 132 851 121 861 l 42 861 l 42 981 l 182 981 b 271 892 244 981 271 954 l 271 349 b 506 114 271 201 364 114 b 740 350 647 114 740 201 l 740 892 b 831 981 740 954 768 981 l 969 981 l 969 861 l 910 861 b 879 831 890 861 879 851 l 879 343 b 507 -17 879 129 731 -17 b 132 343 282 -17 132 129 m 507 1032 b 394 1135 447 1032 394 1071 b 507 1239 394 1200 447 1239 b 619 1135 565 1239 619 1200 b 507 1032 619 1071 565 1032 m 507 1089 b 554 1135 535 1089 554 1108 b 507 1183 554 1164 535 1183 b 460 1135 479 1183 460 1164 b 507 1089 460 1108 479 1089 "},"ů":{"ha":863,"x_min":0,"x_max":0,"o":"m 115 557 b 85 588 115 578 104 588 l 28 588 l 28 703 l 160 703 b 250 617 224 703 250 676 l 250 283 b 381 108 250 186 269 108 b 610 374 525 108 610 235 l 610 703 l 744 703 l 744 146 b 775 115 744 125 756 115 l 832 115 l 832 0 l 703 0 b 614 78 643 0 614 28 l 614 104 b 615 144 614 126 615 144 l 613 144 b 354 -17 583 75 493 -17 b 115 253 196 -17 115 67 m 422 788 b 310 890 363 788 310 826 b 422 994 310 956 363 994 b 535 890 481 994 535 956 b 422 788 535 826 481 788 m 422 844 b 469 890 450 844 469 864 b 422 939 469 919 450 939 b 375 890 394 939 375 919 b 422 844 375 864 394 844 "},"Ű":{"ha":1011,"x_min":0,"x_max":0,"o":"m 132 831 b 101 861 132 851 121 861 l 42 861 l 42 981 l 182 981 b 271 892 244 981 271 954 l 271 349 b 506 114 271 201 364 114 b 740 350 647 114 740 201 l 740 892 b 831 981 740 954 768 981 l 969 981 l 969 861 l 910 861 b 879 831 890 861 879 851 l 879 343 b 507 -17 879 129 731 -17 b 132 343 282 -17 132 129 m 656 1225 l 793 1225 l 660 1049 l 556 1049 m 439 1225 l 556 1225 l 458 1049 l 343 1049 "},"ű":{"ha":863,"x_min":0,"x_max":0,"o":"m 115 557 b 85 588 115 578 104 588 l 28 588 l 28 703 l 160 703 b 250 617 224 703 250 676 l 250 283 b 381 108 250 186 269 108 b 610 374 525 108 610 235 l 610 703 l 744 703 l 744 146 b 775 115 744 125 756 115 l 832 115 l 832 0 l 703 0 b 614 78 643 0 614 28 l 614 104 b 615 144 614 126 615 144 l 613 144 b 354 -17 583 75 493 -17 b 115 253 196 -17 115 67 m 571 981 l 708 981 l 575 804 l 471 804 m 354 981 l 471 981 l 374 804 l 258 804 "},"Ų":{"ha":1011,"x_min":0,"x_max":0,"o":"m 132 831 b 101 861 132 851 121 861 l 42 861 l 42 981 l 182 981 b 271 892 244 981 271 954 l 271 349 b 506 114 271 201 364 114 b 740 350 647 114 740 201 l 740 892 b 831 981 740 954 768 981 l 969 981 l 969 861 l 910 861 b 879 831 890 861 879 851 l 879 338 b 685 18 879 169 783 75 b 565 -124 631 -12 565 -67 b 643 -183 565 -176 643 -183 l 611 -287 b 450 -156 611 -287 450 -275 b 522 -17 450 -97 486 -47 l 522 -14 b 494 -15 522 -14 507 -15 b 132 343 279 -15 132 133 "},"ų":{"ha":863,"x_min":0,"x_max":0,"o":"m 354 -17 b 115 253 196 -17 115 67 l 115 557 b 85 588 115 578 104 588 l 28 588 l 28 703 l 160 703 b 250 588 236 703 250 668 l 250 283 b 381 108 250 186 269 108 b 610 374 525 108 610 235 l 610 703 l 744 703 l 744 146 b 775 115 744 125 756 115 l 832 115 l 832 0 b 743 -124 803 -24 743 -75 b 819 -183 743 -176 819 -183 l 788 -287 b 626 -156 788 -287 626 -275 b 719 -1 626 -86 679 -32 l 719 1 l 693 1 b 614 78 643 1 614 28 l 614 104 b 615 144 614 125 615 144 l 613 144 b 354 -17 613 144 547 -17 "},"Ŵ":{"ha":1344,"x_min":0,"x_max":0,"o":"m 92 831 b 44 861 85 856 69 861 l 24 861 l 24 981 l 103 981 b 224 900 172 981 207 967 l 375 269 b 394 165 388 217 394 165 l 397 165 b 421 269 397 165 406 215 l 618 978 l 743 978 l 926 269 b 947 165 940 215 947 165 l 950 165 b 968 269 950 165 956 217 l 1122 900 b 1242 981 1139 967 1172 981 l 1321 981 l 1321 861 l 1300 861 b 1253 831 1275 861 1260 856 l 1038 0 l 876 0 l 710 619 b 682 754 693 683 682 754 l 679 754 b 647 621 679 754 665 683 l 469 0 l 307 0 m 618 1225 l 742 1225 l 864 1049 l 754 1049 l 682 1160 l 679 1160 l 606 1049 l 497 1049 "},"ŵ":{"ha":1158,"x_min":0,"x_max":0,"o":"m 76 560 b 33 588 69 582 56 588 l 14 588 l 14 703 l 99 703 b 203 628 161 703 185 688 l 328 210 b 344 135 339 174 344 135 l 347 135 b 367 210 347 135 356 174 l 518 700 l 642 700 l 794 210 b 813 135 806 174 813 135 l 815 135 b 832 210 815 135 821 174 l 956 628 b 1061 703 974 688 999 703 l 1144 703 l 1144 588 l 1125 588 b 1082 560 1103 588 1089 582 l 897 0 l 746 0 l 600 451 b 581 528 589 488 581 528 l 578 528 b 558 451 578 528 569 488 l 414 0 l 261 0 m 518 981 l 642 981 l 764 804 l 654 804 l 582 915 l 579 915 l 506 804 l 397 804 "},"Ŷ":{"ha":838,"x_min":0,"x_max":0,"o":"m 350 428 l 101 826 b 42 861 83 856 71 861 l 14 861 l 14 981 l 78 981 b 210 906 144 981 171 971 l 358 658 b 418 549 390 604 418 549 l 421 549 b 481 658 421 549 447 604 l 628 906 b 760 981 667 971 694 981 l 824 981 l 824 861 l 797 861 b 736 826 768 861 754 856 l 489 428 l 489 0 l 350 0 m 358 1225 l 482 1225 l 604 1049 l 494 1049 l 422 1160 l 419 1160 l 346 1049 l 238 1049 "},"ŷ":{"ha":758,"x_min":0,"x_max":0,"o":"m 63 -125 b 165 -175 63 -126 106 -175 b 289 -74 219 -175 263 -136 l 321 3 l 81 560 b 36 588 71 582 58 588 l 19 588 l 19 703 l 99 703 b 199 632 156 703 176 688 l 358 236 b 383 151 372 200 383 151 l 386 151 b 410 236 386 151 396 200 l 556 632 b 658 703 576 688 601 703 l 739 703 l 739 588 l 721 588 b 676 560 699 588 686 582 l 399 -126 b 172 -292 356 -235 271 -292 b 7 -222 68 -292 7 -222 l 63 -125 m 317 981 l 440 981 l 563 804 l 453 804 l 381 915 l 378 915 l 304 804 l 196 804 "},"Ÿ":{"ha":838,"x_min":0,"x_max":0,"o":"m 350 428 l 101 826 b 42 861 83 856 71 861 l 14 861 l 14 981 l 78 981 b 210 906 144 981 171 971 l 358 658 b 418 549 390 604 418 549 l 421 549 b 481 658 421 549 447 604 l 628 906 b 760 981 667 971 694 981 l 824 981 l 824 861 l 797 861 b 736 826 768 861 754 856 l 489 428 l 489 0 l 350 0 m 485 1225 l 586 1225 l 586 1085 l 485 1085 m 256 1225 l 358 1225 l 358 1085 l 256 1085 "},"Ź":{"ha":808,"x_min":0,"x_max":0,"o":"m 25 92 l 508 772 b 582 863 547 826 582 863 l 582 865 b 508 861 582 865 556 861 l 193 861 b 163 831 174 861 163 851 l 163 768 l 36 768 l 36 892 b 125 981 36 954 63 981 l 754 981 l 754 890 l 269 208 b 197 119 232 154 197 119 l 197 117 b 269 119 197 117 224 119 l 615 119 b 646 150 635 119 646 131 l 646 213 l 774 213 l 774 89 b 685 0 774 26 747 0 l 25 0 m 458 1225 l 594 1225 l 463 1049 l 357 1049 "},"ź":{"ha":701,"x_min":0,"x_max":0,"o":"m 46 78 l 406 522 b 464 588 436 560 464 588 l 464 590 b 396 588 464 590 442 588 l 204 588 b 174 557 185 588 174 576 l 174 510 l 50 510 l 50 614 b 139 703 50 676 76 703 l 642 703 l 642 625 l 281 179 b 222 115 251 144 222 115 l 222 113 b 290 115 222 113 244 115 l 515 115 b 546 146 535 115 546 126 l 546 193 l 668 193 l 668 89 b 579 0 668 26 642 0 l 46 0 m 408 981 l 544 981 l 413 804 l 307 804 "},"Ż":{"ha":808,"x_min":0,"x_max":0,"o":"m 25 92 l 508 772 b 582 863 547 826 582 863 l 582 865 b 508 861 582 865 556 861 l 193 861 b 163 831 174 861 163 851 l 163 768 l 36 768 l 36 892 b 125 981 36 954 63 981 l 754 981 l 754 890 l 269 208 b 197 119 232 154 197 119 l 197 117 b 269 119 197 117 224 119 l 615 119 b 646 150 635 119 646 131 l 646 213 l 774 213 l 774 89 b 685 0 774 26 747 0 l 25 0 m 351 1225 l 472 1225 l 472 1085 l 351 1085 "},"ż":{"ha":701,"x_min":0,"x_max":0,"o":"m 46 78 l 406 522 b 464 588 436 560 464 588 l 464 590 b 396 588 464 590 442 588 l 204 588 b 174 557 185 588 174 576 l 174 510 l 50 510 l 50 614 b 139 703 50 676 76 703 l 642 703 l 642 625 l 281 179 b 222 115 251 144 222 115 l 222 113 b 290 115 222 113 244 115 l 515 115 b 546 146 535 115 546 126 l 546 193 l 668 193 l 668 89 b 579 0 668 26 642 0 l 46 0 m 301 981 l 422 981 l 422 840 l 301 840 "},"Ž":{"ha":808,"x_min":0,"x_max":0,"o":"m 25 92 l 508 772 b 582 863 547 826 582 863 l 582 865 b 508 861 582 865 556 861 l 193 861 b 163 831 174 861 163 851 l 163 768 l 36 768 l 36 892 b 125 981 36 954 63 981 l 754 981 l 754 890 l 269 208 b 197 119 232 154 197 119 l 197 117 b 269 119 197 117 224 119 l 615 119 b 646 150 635 119 646 131 l 646 213 l 774 213 l 774 89 b 685 0 774 26 747 0 l 25 0 m 228 1225 l 338 1225 l 410 1113 l 413 1113 l 485 1225 l 596 1225 l 474 1049 l 350 1049 "},"ž":{"ha":701,"x_min":0,"x_max":0,"o":"m 46 78 l 406 522 b 464 588 436 560 464 588 l 464 590 b 396 588 464 590 442 588 l 204 588 b 174 557 185 588 174 576 l 174 510 l 50 510 l 50 614 b 139 703 50 676 76 703 l 642 703 l 642 625 l 281 179 b 222 115 251 144 222 115 l 222 113 b 290 115 222 113 244 115 l 515 115 b 546 146 535 115 546 126 l 546 193 l 668 193 l 668 89 b 579 0 668 26 642 0 l 46 0 m 178 981 l 288 981 l 360 868 l 363 868 l 435 981 l 546 981 l 424 804 l 300 804 "},"ſ":{"ha":451,"x_min":0,"x_max":0,"o":"m 118 594 l 31 594 l 31 703 l 118 703 l 118 726 b 383 988 118 958 294 988 b 435 983 414 988 435 983 l 435 868 b 400 871 435 868 421 871 b 253 722 347 871 253 853 l 253 0 l 118 0 "},"ƒ":{"ha":500,"x_min":0,"x_max":0,"o":"m -24 -78 b -74 -75 -53 -78 -74 -75 l -62 42 b -29 39 -62 42 -49 39 b 131 193 22 39 118 57 l 156 486 l 51 486 l 51 594 l 165 594 l 178 726 b 464 988 199 958 375 988 b 515 983 496 988 515 983 l 504 868 b 471 871 504 868 492 871 b 311 717 418 871 324 853 l 299 594 l 446 594 l 446 486 l 290 486 l 264 183 b -24 -78 243 -49 65 -78 "},"Ș":{"ha":740,"x_min":0,"x_max":0,"o":"m 129 228 b 375 110 129 228 229 110 b 543 256 467 110 543 165 b 69 735 543 461 69 429 b 383 997 69 882 199 997 b 663 817 488 997 663 950 l 663 736 l 535 736 l 535 782 b 383 871 535 833 461 871 b 211 739 278 871 211 813 b 683 263 211 540 683 585 b 372 -17 683 113 567 -17 b 47 129 164 -17 47 129 m 426 -82 b 435 -150 432 -107 435 -129 b 401 -278 435 -193 424 -231 l 306 -278 b 332 -164 322 -235 332 -200 b 317 -82 332 -139 328 -112 "},"ș":{"ha":631,"x_min":0,"x_max":0,"o":"m 107 197 b 319 93 107 197 190 93 b 442 183 385 93 442 122 b 58 522 442 310 58 296 b 319 719 58 657 176 719 b 550 583 404 719 550 690 l 550 517 l 429 517 l 429 551 b 324 611 429 593 367 611 b 193 528 244 611 193 583 b 576 189 193 394 576 424 b 318 -17 576 64 465 -17 b 38 107 131 -17 38 107 m 367 -82 b 375 -150 372 -107 375 -129 b 342 -278 375 -193 364 -231 l 246 -278 b 272 -164 263 -235 272 -200 b 257 -82 272 -139 268 -112 "},"Ț":{"ha":883,"x_min":0,"x_max":0,"o":"m 372 861 l 175 861 b 144 831 156 861 144 851 l 144 768 l 21 768 l 21 897 b 104 981 21 960 42 981 l 779 981 b 863 897 842 981 863 960 l 863 768 l 739 768 l 739 831 b 708 861 739 851 728 861 l 511 861 l 511 0 l 372 0 m 492 -82 b 500 -150 497 -107 500 -129 b 467 -278 500 -193 489 -231 l 371 -278 b 397 -164 388 -235 397 -200 b 382 -82 397 -139 393 -112 "},"ț":{"ha":489,"x_min":0,"x_max":0,"o":"m 124 594 l 32 594 l 32 703 l 126 703 l 126 896 l 258 896 l 258 703 l 426 703 l 426 594 l 258 594 l 258 272 b 407 114 258 132 356 114 b 439 117 426 114 439 117 l 439 -3 b 392 -6 439 -3 421 -6 b 124 257 303 -6 124 22 m 329 -82 b 338 -150 335 -107 338 -129 b 304 -278 338 -193 326 -231 l 208 -278 b 235 -164 225 -235 235 -200 b 219 -82 235 -139 231 -112 "},"ˆ":{"ha":674,"x_min":0,"x_max":0,"o":"m 275 1225 l 399 1225 l 521 1049 l 411 1049 l 339 1160 l 336 1160 l 263 1049 l 154 1049 "},"ˇ":{"ha":678,"x_min":0,"x_max":0,"o":"m 153 1225 l 263 1225 l 335 1113 l 338 1113 l 410 1225 l 521 1225 l 399 1049 l 275 1049 "},"˘":{"ha":696,"x_min":0,"x_max":0,"o":"m 338 1043 b 160 1206 250 1043 160 1099 l 160 1225 l 253 1225 l 253 1207 b 338 1124 253 1149 294 1124 b 424 1208 381 1124 424 1149 l 424 1225 l 515 1225 l 515 1206 b 338 1043 515 1099 426 1043 "},"˙":{"ha":356,"x_min":0,"x_max":0,"o":"m 117 1225 l 238 1225 l 238 1085 l 117 1085 "},"˚":{"ha":528,"x_min":0,"x_max":0,"o":"m 264 1032 b 151 1135 204 1032 151 1071 b 264 1239 151 1200 204 1239 b 376 1135 322 1239 376 1200 b 264 1032 376 1071 322 1032 m 264 1089 b 311 1135 292 1089 311 1108 b 264 1183 311 1164 292 1183 b 217 1135 236 1183 217 1164 b 264 1089 217 1108 236 1089 "},"˛":{"ha":501,"x_min":0,"x_max":0,"o":"m 150 -156 b 276 24 150 -50 276 24 l 372 14 b 265 -124 372 14 265 -61 b 343 -183 265 -176 343 -183 l 311 -287 b 150 -156 311 -287 150 -275 "},"˜":{"ha":721,"x_min":0,"x_max":0,"o":"m 128 1050 b 271 1226 128 1167 176 1226 b 449 1132 367 1226 388 1132 b 504 1222 490 1132 504 1182 l 596 1222 b 453 1046 596 1107 547 1046 b 275 1140 358 1046 336 1140 b 219 1050 233 1140 219 1092 l 128 1050 "},"˝":{"ha":743,"x_min":0,"x_max":0,"o":"m 450 1225 l 588 1225 l 454 1049 l 350 1049 m 233 1225 l 350 1225 l 253 1049 l 138 1049 "},"ẞ":{"ha":889,"x_min":0,"x_max":0,"o":"m 42 119 l 101 119 b 132 150 121 119 132 131 l 132 703 b 414 981 132 864 250 981 l 792 981 l 792 888 l 599 606 b 565 560 579 576 565 560 l 565 557 l 579 557 b 833 285 732 557 833 418 b 543 0 833 126 703 0 l 411 0 l 411 119 l 533 119 b 692 285 629 119 692 197 b 535 447 692 372 631 447 l 414 447 l 414 543 l 571 778 b 632 863 613 840 632 863 l 632 865 b 572 861 632 865 622 861 l 418 861 b 271 706 332 861 271 793 l 271 89 b 182 0 271 26 244 0 l 42 0 "},"–":{"ha":906,"x_min":0,"x_max":0,"o":"m 88 450 l 818 450 l 818 342 l 88 342 "},"—":{"ha":1183,"x_min":0,"x_max":0,"o":"m 88 450 l 1096 450 l 1096 342 l 88 342 "},"‘":{"ha":317,"x_min":0,"x_max":0,"o":"m 156 996 l 251 996 l 192 725 l 65 725 "},"’":{"ha":317,"x_min":0,"x_max":0,"o":"m 125 997 l 251 997 l 161 726 l 65 726 "},"‚":{"ha":317,"x_min":0,"x_max":0,"o":"m 251 146 l 161 -125 l 65 -125 l 125 146 "},"“":{"ha":510,"x_min":0,"x_max":0,"o":"m 347 996 l 444 996 l 385 725 l 257 725 m 156 996 l 251 996 l 192 725 l 65 725 "},"”":{"ha":510,"x_min":0,"x_max":0,"o":"m 317 997 l 444 997 l 354 726 l 257 726 m 125 997 l 251 997 l 161 726 l 65 726 "},"„":{"ha":510,"x_min":0,"x_max":0,"o":"m 317 146 l 444 146 l 354 -125 l 257 -125 m 251 146 l 161 -125 l 65 -125 l 125 146 "},"†":{"ha":547,"x_min":0,"x_max":0,"o":"m 207 596 l 39 596 l 39 703 l 207 703 l 207 981 l 336 981 l 336 703 l 508 703 l 508 596 l 336 596 l 336 -69 l 207 -69 "},"‡":{"ha":547,"x_min":0,"x_max":0,"o":"m 207 238 l 39 238 l 39 346 l 207 346 l 207 596 l 39 596 l 39 703 l 207 703 l 207 981 l 336 981 l 336 703 l 508 703 l 508 596 l 336 596 l 336 346 l 508 346 l 508 238 l 336 238 l 336 -69 l 207 -69 "},"•":{"ha":631,"x_min":0,"x_max":0,"o":"m 315 189 b 100 404 196 189 100 285 b 315 619 100 524 196 619 b 529 404 433 619 529 524 b 315 189 529 285 433 189 "},"…":{"ha":1256,"x_min":0,"x_max":0,"o":"m 919 143 l 1057 143 l 1057 0 l 919 0 m 558 143 l 697 143 l 697 0 l 558 0 m 199 143 l 336 143 l 336 0 l 199 0 "},"‰":{"ha":1546,"x_min":0,"x_max":0,"o":"m 292 586 b 82 792 175 586 82 678 b 292 997 82 904 175 997 b 503 792 408 997 503 904 b 292 586 503 678 408 586 m 853 981 l 983 981 l 233 0 l 103 0 m 292 689 b 394 792 350 689 394 733 b 292 896 394 849 350 896 b 189 792 233 896 189 849 b 292 689 189 733 233 689 m 1078 189 b 1288 394 1078 301 1171 394 b 1499 189 1404 394 1499 301 b 1288 -17 1499 75 1404 -17 b 1078 189 1171 -17 1078 75 m 585 189 b 796 394 585 301 678 394 b 1007 189 911 394 1007 301 b 796 -17 1007 75 911 -17 b 585 189 678 -17 585 75 m 1185 189 b 1288 86 1185 131 1229 86 b 1390 189 1346 86 1390 131 b 1288 293 1390 246 1346 293 b 1185 189 1229 293 1185 246 m 693 189 b 796 86 693 131 738 86 b 899 189 853 86 899 131 b 796 293 899 246 853 293 b 693 189 738 293 693 246 "},"‹":{"ha":510,"x_min":0,"x_max":0,"o":"m 58 392 l 286 678 l 421 678 l 193 392 l 421 107 l 286 107 "},"›":{"ha":510,"x_min":0,"x_max":0,"o":"m 314 392 l 86 678 l 221 678 l 449 392 l 221 107 l 86 107 "},"⁄":{"ha":581,"x_min":0,"x_max":0,"o":"m 435 981 l 546 981 l 147 0 l 36 0 "},"€":{"ha":835,"x_min":0,"x_max":0,"o":"m 54 446 l 136 446 b 133 492 136 446 133 471 b 136 546 133 513 136 546 l 54 546 l 54 635 l 151 635 b 622 997 208 846 393 997 b 756 981 696 997 756 981 l 722 854 b 619 869 722 854 674 869 b 299 635 464 869 344 776 l 683 635 l 664 546 l 278 546 b 275 494 278 546 275 522 b 278 446 275 478 278 446 l 646 446 l 629 358 l 301 358 b 624 114 344 213 468 114 b 753 133 696 114 753 133 l 781 7 b 622 -17 781 7 713 -17 b 150 358 385 -17 203 138 l 54 358 "},"™":{"ha":1365,"x_min":0,"x_max":0,"o":"m 588 463 l 619 463 b 635 475 629 463 633 465 l 675 981 l 771 981 l 915 661 b 935 603 926 636 935 603 l 938 603 b 956 661 938 603 946 636 l 1100 981 l 1197 981 l 1238 475 b 1251 463 1238 465 1242 463 l 1285 463 l 1285 369 l 1203 369 b 1135 432 1153 369 1139 382 l 1111 735 b 1113 776 1110 756 1113 776 l 1110 776 l 978 482 l 893 482 l 761 776 l 758 776 b 760 735 758 776 761 756 l 736 432 b 668 369 732 382 718 369 l 588 369 m 265 888 l 160 888 b 147 875 150 888 147 883 l 147 833 l 54 833 l 54 918 b 119 981 54 967 68 981 l 517 981 b 582 918 567 981 582 967 l 582 833 l 490 833 l 490 875 b 478 888 490 883 488 888 l 371 888 l 371 369 l 265 369 "},"−":{"ha":851,"x_min":0,"x_max":0,"o":"m 67 450 l 785 450 l 785 342 l 67 342 "},"∞":{"ha":706,"x_min":0,"x_max":0,"o":"m 353 232 b 100 485 214 232 100 346 b 353 736 100 625 214 736 b 606 485 493 736 606 625 b 353 232 606 346 493 232 "},"≈":{"ha":851,"x_min":0,"x_max":0,"o":"m 79 528 b 289 629 126 578 204 629 b 564 532 419 629 465 532 b 717 608 621 532 679 568 l 774 526 b 564 426 726 478 649 426 b 289 524 433 426 388 524 b 136 446 231 524 174 485 m 79 265 b 289 367 126 315 204 367 b 564 269 419 367 465 269 b 717 346 621 269 679 306 l 774 264 b 564 164 726 215 649 164 b 289 261 433 164 388 261 b 136 185 231 261 174 224 "},"≠":{"ha":851,"x_min":0,"x_max":0,"o":"m 247 210 l 94 210 l 94 318 l 319 318 l 431 472 l 94 472 l 94 581 l 503 581 l 617 740 l 688 692 l 607 581 l 757 581 l 757 472 l 535 472 l 425 318 l 757 318 l 757 210 l 351 210 l 242 54 l 169 101 "},"≤":{"ha":851,"x_min":0,"x_max":0,"o":"m 93 442 l 740 731 l 740 607 l 243 397 l 243 394 l 740 185 l 740 61 l 93 350 m 740 -132 l 740 -240 l 94 -240 l 94 -132 "},"≥":{"ha":851,"x_min":0,"x_max":0,"o":"m 111 185 l 608 394 l 608 397 l 111 607 l 111 731 l 760 442 l 760 350 l 111 61 m 757 -132 l 757 -240 l 111 -240 l 111 -132 "},"ﬀ":{"ha":888,"x_min":0,"x_max":0,"o":"m 118 594 l 31 594 l 31 703 l 118 703 l 118 726 b 383 988 118 958 294 988 b 435 983 414 988 435 983 l 435 868 b 400 871 435 868 421 871 b 253 722 347 871 253 853 l 253 703 l 556 703 l 556 726 b 821 988 556 958 732 988 b 872 983 851 988 872 983 l 872 868 b 838 871 872 868 857 871 b 690 722 785 871 690 853 l 690 703 l 856 703 l 856 594 l 690 594 l 690 0 l 556 0 l 556 594 l 253 594 l 253 0 l 118 0 "},"ﬁ":{"ha":819,"x_min":0,"x_max":0,"o":"m 118 594 l 31 594 l 31 703 l 118 703 l 118 726 b 383 988 118 958 294 988 b 435 983 414 988 435 983 l 435 868 b 397 871 435 868 419 871 b 253 722 344 871 253 853 l 253 703 l 607 703 b 696 614 669 703 696 676 l 696 146 b 726 115 696 125 707 115 l 783 115 l 783 0 l 651 0 b 561 89 588 0 561 26 l 561 564 b 531 594 561 583 550 594 l 253 594 l 253 0 l 118 0 m 564 981 l 686 981 l 686 840 l 564 840 "},"ﬂ":{"ha":810,"x_min":0,"x_max":0,"o":"m 599 981 b 688 892 661 981 688 954 l 688 146 b 718 115 688 125 699 115 l 775 115 l 775 0 l 642 0 b 553 89 579 0 553 26 l 553 835 b 522 865 553 856 542 865 l 400 865 b 253 724 347 865 253 854 l 253 703 l 419 703 l 419 594 l 253 594 l 253 0 l 118 0 l 118 594 l 31 594 l 31 703 l 118 703 l 118 726 b 383 981 118 958 294 981 "},"ﬃ":{"ha":1256,"x_min":0,"x_max":0,"o":"m 118 594 l 31 594 l 31 703 l 118 703 l 118 726 b 383 988 118 958 294 988 b 435 983 414 988 435 983 l 435 868 b 397 871 435 868 419 871 b 253 722 346 871 253 853 l 253 703 l 556 703 l 556 726 b 821 988 556 958 732 988 b 872 983 851 988 872 983 l 872 868 b 835 871 872 868 857 871 b 690 722 782 871 690 853 l 690 703 l 1044 703 b 1133 614 1107 703 1133 676 l 1133 146 b 1164 115 1133 125 1144 115 l 1221 115 l 1221 0 l 1088 0 b 999 89 1025 0 999 26 l 999 564 b 968 594 999 583 988 594 l 690 594 l 690 0 l 556 0 l 556 594 l 253 594 l 253 0 l 118 0 m 1001 981 l 1122 981 l 1122 840 l 1001 840 "},"ﬄ":{"ha":1246,"x_min":0,"x_max":0,"o":"m 1035 981 b 1124 892 1097 981 1124 954 l 1124 146 b 1154 115 1124 125 1135 115 l 1211 115 l 1211 0 l 1078 0 b 989 89 1015 0 989 26 l 989 835 b 958 865 989 856 978 865 l 836 865 b 689 724 783 865 689 854 l 689 703 l 856 703 l 856 594 l 689 594 l 689 0 l 554 0 l 554 594 l 253 594 l 253 0 l 118 0 l 118 594 l 31 594 l 31 703 l 118 703 l 118 729 b 381 988 118 958 290 988 b 435 983 413 988 435 983 l 435 868 b 397 871 435 868 418 871 b 253 724 344 871 253 851 l 253 703 l 554 703 l 554 726 b 819 981 554 958 731 981 "}},"familyName":"Museo 500","ascender":1306,"descender":-361,"underlinePosition":-75,"underlineThickness":50,"boundingBox":{"yMin":-237,"xMin":-53,"yMax":892,"xMax":1124},"resolution":1000,"original_font_information":{"format":0,"copyright":"Copyright (c) 2008 by Jos Buivenga/exljbris. All rights reserved.","fontFamily":"Museo 500","fontSubfamily":"Regular","uniqueID":"FONTLAB:OTFEXPORT","fullName":"Museo500-Regular","version":"1.071","postScriptName":"Museo500-Regular","trademark":"Museo is a trademark of Jos Buivenga/exljbris.","manufacturer":"Jos Buivenga","designer":"Jos Buivenga","description":"Copyright (c) 2008 by Jos Buivenga/exljbris. All rights reserved.","manufacturerURL":"http://www.josbuivenga.demon.nl","designerURL":"http://www.josbuivenga.demon.nl","preferredFamily":"Museo","preferredSubfamily":"500"},"cssFontWeight":"normal","cssFontStyle":"normal"};
-module.exports = fontDef;
