+++ src/components/animation-checkbox.tsx
