+++ src/components/canvas-view.tsx
