+++ src/components/day-slider.tsx
