+++ src/components/earth-view-comp.tsx
