+++ src/components/rays-view-comp.tsx
