+++ src/components/slider.tsx
