+++ src/css/jquery-ui-theme.less
@@ -1,7 +1,7 @@
 .grasp-seasons {
   // Note that jquery-ui theme imports are scoped to .grasp-seasons, so we won't pollute rest of the page.
   // Use (less) to enforce LESS processing. Otherwise, style won't respect scope nesting.
-  @import (less) "../node_modules/jquery-ui/themes/base/slider.css";
+  @import (less) "../../node_modules/jquery-ui/themes/base/slider.css";
 
   .ui-widget {
     font-family: inherit;
