+++ src/global.d.ts
