+++ src/scripting-api.ts
