+++ src/views/earth-view-interaction.ts
