+++ src/views/earth-view.ts
